% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addDuToitCounts}
\alias{addDuToitCounts}
\title{Add DuToit's Pan data from Doug}
\usage{
addDuToitCounts(counts, config)
}
\arguments{
\item{counts}{A dataframe with CWAC data coming out of \code{\link[CWAC]{getCwacSppCounts}} SAFRING reference number of the species we want to add data to.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}.}
}
\value{
A dataframe with DuToit's counts added to \code{counts}
}
\description{
This function is used to attach data from DuToit's Pan that
Doug Harebottle provided and that are not on the CWAC database.
}
\examples{
\dontrun{
config <- configPipeline(
    year = 2021,
    dur = 29,
    mod_file = "cwac_ssm_two_season_mean_rev.R",
    package = "jagsUI",
    data_dir = NULL,
    out_dir = NULL,
    server = FALSE
)

site_code <- 28462448

counts <- CWAC::getCwacSiteCounts(site_code)

complete_counts <- addDuToitCounts(counts, config)


}
}
