% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOccSiteCovts.R
\name{addOccSiteCovt}
\alias{addOccSiteCovt}
\title{Add occupancy site covariate}
\usage{
addOccSiteCovt(sites, covt, years, covts_dir, file_fix)
}
\arguments{
\item{sites}{A spatial object with the sites where data was collected.}

\item{covt}{A character string with the names of the covariate to be
extracted. The names must correspond to the names contained the covariate
file name (e.g. "prcp", "tmax", etc.). At the moment the function assumes
that covariates are found at: "covts_dir_file_fix\link{1}covt_file_fix\link{2}.rds"}

\item{years}{A vector of years that we want to extract}

\item{covts_dir}{Directory where covariate are found.}

\item{file_fix}{A character vector with two elements corresponding to
additional characters that are found before and after covts.}
}
\description{
Add occupancy site covariate
}
