% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOccVisitCovt.R
\name{addOccVisitCovt}
\alias{addOccVisitCovt}
\title{Add occupancy visit covariate}
\usage{
addOccVisitCovt(visits, sites, covt, covts_dir, file_fix)
}
\arguments{
\item{visits}{A dataframe with detection/non-detection data. It must contain
at least the following columns: "year", "month", "SiteName".}

\item{sites}{A spatial object with the sites where data was collected.}

\item{covt}{A character string with the names of the covariate to be
extracted. The names must correspond to the names contained the covariate
file name (e.g. "prcp", "tmax", etc.). At the moment the function assumes
that covariates are found at: "covts_dir_file_fix\link{1}covt_file_fix\link{2}.rds"}

\item{covts_dir}{Directory where covariate are found.}

\item{file_fix}{A character vector with two elements corresponding to
additional characters that are found before and after covts.}
}
\value{
The sites dataframe with an additional column corresponding to
the new variable.
}
\description{
Add occupancy visit covariate
}
