% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOccVisitCovts.R
\name{addOccVisitCovts}
\alias{addOccVisitCovts}
\title{Add occupancy visit covariates}
\usage{
addOccVisitCovts(visit_data, sites, covts, covts_dir, file_fix)
}
\arguments{
\item{visit_data}{A dataframe with occupancy data}

\item{sites}{A spatial object with the sites where visit_data was collected.}

\item{covts}{A character vector with the names of the covariates to be
extracted. The names must correspond to the names contained the covariate
file name (e.g. "prcp", "tmax", etc.). At the moment the function assumes
that covariates are found at: "covts_dir_file_fix\link{1}covts_file_fix\link{2}.rds"}

\item{covts_dir}{Directory where covariates are found.}

\item{file_fix}{A character vector with two elements correspondint to
additional characters that are found before and after covts.}
}
\description{
Add occupancy visit covariates
}
