% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gee.R
\name{addSiteVarEETimeLimit}
\alias{addSiteVarEETimeLimit}
\title{Annotate site data with Google Earth Engine with a time limit}
\usage{
addSiteVarEETimeLimit(
  ee_feats,
  ee_collection,
  band,
  last_year,
  reducer,
  unmask,
  monitor,
  config
)
}
\arguments{
\item{ee_feats}{A feature collection with the sites we want to annotate.
We should have uploaded an sf object with the sites to GEE, previously.
See \code{\link{uploadFeaturesToEE}}}

\item{ee_collection}{A GEE collection produced with \code{ee$ImageCollection()}.
See \href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{band}{The band of the collection we are going to use to annotate our data}

\item{last_year}{The last year provided by the GEE image collection}

\item{reducer}{The reducer we will use to summarize the images of the image collection}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}

\item{monitor}{Logical. If TRUE (default) monitoring messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}
}
\description{
Annotate site data with Google Earth Engine with a time limit
}
\details{
The function annotates data with the corresponding year if it is present in the
GEE image collection. Years after the last year present in 'config$years' are annotated
with the last year present in the image collection.
}
