% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barberspan.R
\docType{data}
\name{barberspan}
\alias{barberspan}
\title{Barberspan CWAC data}
\format{
A data frame with 1579 rows and 34 variables:
\describe{
\item{item}{DESCRIPTION}
...
}
}
\usage{
barberspan
}
\description{
A dataset containing all counts recorded for Barberspan in the CWAC project
}
\keyword{datasets}
