% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{birdie-class}
\alias{birdie-class}
\alias{coef.birdie}
\alias{fitted.birdie}
\alias{predict.birdie}
\alias{simulate.birdie}
\alias{print.birdie}
\alias{summary.birdie}
\title{Class "birdie_fit" of BIRDiE Models}
\usage{
\method{coef}{birdie}(object, complete = FALSE, ...)

\method{fitted}{birdie}(object, ...)

\method{predict}{birdie}(object, ...)

\method{simulate}{birdie}(object, nsim = 1, seed = NULL, ...)

\method{print}{birdie}(x, ...)

\method{summary}{birdie}(object, ...)
}
\arguments{
\item{object, x}{A \code{birdie} object}

\item{complete}{If \code{TRUE}, return group-level (rather than marginal)
coefficient estimates as a 3D array.}

\item{...}{Potentially further arguments passed from other methods}

\item{nsim}{The number of vectors to simulate. Defaults to 1.}

\item{seed}{Used to seed the random number generator. See \code{\link[stats:simulate]{stats::simulate()}}.}
}
\description{
The output of \code{\link[=birdie]{birdie()}} is an object of class \code{birdie}, which supports
many generic functions. Notably \code{coef.birdie()} returns the main model
estimates of outcome given race, and \code{fitted.birdie()} returns a table
analogous to the output of \code{\link[=bisg]{bisg()}} with updated race probabilities.
}
\details{
The internal structure of \code{birdie} objects is not designed to be accessed
directly. The generics listed here should be used instead.
}
\section{Functions}{
\itemize{
\item \code{coef(birdie)}: Return estimated outcome-given-race distributions.

\item \code{fitted(birdie)}: Return an updated race probability table. \code{\link[=bisg]{bisg()}}
estimates \code{Pr(R | G, X, S)}; this table estimates \code{Pr(R | Y, G, X, S)}.

\item \code{predict(birdie)}: Alias for \code{fitted.birdie}.

\item \code{simulate(birdie)}: Simulate race from the posterior distribution
\code{Pr(R | Y, G, X, S, Theta-MAP)}. Does not account for uncertainty in model
parameters.

\item \code{print(birdie)}: Print a summary of the model fit.

\item \code{summary(birdie)}: Print a summary of the model fit.

}}
