% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_helpers.R
\name{birdie-family}
\alias{birdie-family}
\alias{cat_dir}
\alias{cat_mixed}
\title{BIRDiE Complete-Data Model Families}
\usage{
cat_dir(link = "identity")

cat_mixed(link = "softmax")
}
\arguments{
\item{link}{The link function. Only one option available for categorical
regression models.}
}
\value{
A list of class `family` containing the specification.
}
\description{
BIRDiE supports a number of complete-data outcome models, including
categorical regression models. Models specific to BIRDiE are listed here.
See the Details section of [birdie()] for more information about each
model.
}
\examples{
cat_dir()
cat_mixed()

}
