% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{calc_joint_bisgz}
\alias{calc_joint_bisgz}
\title{Estimate the joint distribution of R and X}
\usage{
calc_joint_bisgz(
  r_probs,
  x,
  method = c("weight", "thresh", "mi", "ols"),
  prefix = "pr_",
  n_mi = 8
)
}
\arguments{
\item{r_probs}{a matrix data frame of race probabilities}

\item{x}{the variable X to tabulate against}

\item{method}{the estimation method: a weighted-mean estimator, a
thresholding estimator, or a multiple-imputation estimator.}

\item{prefix}{how to select the race probability columns from \code{r_probs}, if
if is a data frame}

\item{n_mi}{if \code{method == "mi"}, how many imputations to produce.}
}
\value{
a matrix
}
\description{
Estimate the joint distribution of R and X
}
