% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{calc_joint_bisgz_ols}
\alias{calc_joint_bisgz_ols}
\title{Estimate the joint distribution of R and X}
\usage{
calc_joint_bisgz_ols(r_probs, x, p_GZ, prefix = "pr_", n_mi = 8)
}
\arguments{
\item{r_probs}{a matrix data frame of race probabilities}

\item{x}{the variable X to tabulate against}

\item{prefix}{how to select the race probability columns from \code{r_probs}, if
if is a data frame}

\item{n_mi}{if \code{method == "mi"}, how many imputations to produce.}

\item{method}{the estimation method: a weighted-mean estimator, a
thresholding estimator, or a multiple-imputation estimator.}
}
\value{
a matrix
}
\description{
Estimate the joint distribution of R and X
}
