% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{calc_joint_model}
\alias{calc_joint_model}
\title{Calculate a posterior quantile of the joint distribution of R and X}
\usage{
calc_joint_model(
  fit,
  which = "global",
  q = 0.5,
  p_r = c(white = 0.63, black = 0.121, hisp = 0.173, asian = 0.0478, aian = 0.0072, other
    = 0.021)
)
}
\arguments{
\item{fit}{a \code{fit_birdie} object (the output of \code{model_race})}

\item{which}{if \code{condition} was used in \code{model_race}, which estimates to extract.}

\item{q}{the quantile}

\item{p_r}{a vector containing the marginal probabilities for each value of
\code{R}. Defaults to the demographics of the US.}
}
\value{
a matrix
}
\description{
Calculate a posterior quantile of the joint distribution of R and X
}
