% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkListEqual.R
\name{checkListEqual}
\alias{checkListEqual}
\title{Checks that all elements in a list are identical}
\usage{
checkListEqual(x)
}
\arguments{
\item{x}{A list}
}
\value{
TRUE if all elements in a list are identical, FALSE otherwise. If the function finds an element that is different it will print the elements that differ
}
\description{
Checks that all elements in a list are identical
}
\examples{
checkListEqual(x = list(a = 1:5, b = 3:6))
checkListEqual(x = list(a = letters[7:9], b = c("g", "h", "i")))
}
