% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineAbuDiags}
\alias{combineAbuDiags}
\title{Combine species abundance diagnostics into a single object}
\usage{
combineAbuDiags(config, sp_codes, year)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}.}

\item{sp_codes}{SAFRING reference numbers of the species we want diagnostics for.}

\item{year}{The year(s) for which diagnostics are required. Years should be formated
as the two last digits of the beginning and last years. See}
}
\value{
A dataframe with diagnostic paramters
}
\description{
The data pipeline produces one abundance model diagnostic file for each
species and year. This function reads diagnostics files and combines them into
a single object
}
\examples{
\dontrun{
config <- configPipeline(
    year = 2021,
    dur = 29,
    mod_file = "cwac_ssm_two_season_mean_rev.R",
    package = "jagsUI",
    data_dir = NULL,
    out_dir = NULL,
    server = FALSE
)
sp_codes <- config$species

combineAbuDiags(config, sp_codes, "93_21")
}
}
