% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineOccuDiags}
\alias{combineOccuDiags}
\title{Combine species occupancy diagnostics into a single object}
\usage{
combineOccuDiags(config, sp_codes, year)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}.}

\item{sp_codes}{SAFRING reference numbers of the species we want diagnostics for.}

\item{year}{The year for which diagnostics are required.}
}
\value{
A dataframe with diagnostic paramters
}
\description{
The datapipeline produces one occupancy diagnostic file for each
species and year. This function reads diagnostics files and combines them into
a single object
}
\examples{
\dontrun{
config <- configPipeline(year = 2010,
                         dur = 3,
                         occ_mod = c("log_dist_coast", "elev", "log_hum.km2", "wetcon",
                                     "watrec", "watext", "log_watext", "watext:watrec",
                                     "ndvi", "prcp", "tdiff"),
                         det_mod = c("(1|obs_id)", "log_hours", "prcp", "tdiff", "cwac"),
                         fixed_vars = c("Pentad", "lon", "lat", "watocc_ever", "wetext_2018","wetcon_2018",
                                        "dist_coast", "elev"),
                         package = "spOccupancy",
                         data_dir = "analysis/hpc/imports",
                         out_dir = "analysis/hpc/imports",
                         server = TRUE)
sp_codes <- config$species

combineOccuDiags(config, sp_codes, 2008)
}
}
