% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPipeline.R
\name{configPipeline}
\alias{configPipeline}
\title{Configure pipeline parameters}
\usage{
configPipeline(
  year,
  dur,
  region = c("southafrica", "kenya"),
  module = c("dst", "abu"),
  occ_mod = NULL,
  det_mod = NULL,
  fixed_vars = NULL,
  mod_file = NULL,
  server = FALSE,
  data_dir = NULL,
  mod_dir = NULL,
  out_dir = NULL,
  package = NULL
)
}
\arguments{
\item{year}{Year of interest.}

\item{dur}{Temporal coverage of the analysis in years. \code{year} will be the last year
covered by the analysis.}

\item{region}{A character string with the region we want to run the pipeline for.
Currently only "South Africa" and "Kenya" are supported.}

\item{module}{A character string defining the module the pipeline should run.
At the moment it can be one of \code{c("dst", "abu")} for distributions and abundance
respectively.}

\item{occ_mod}{A character vector with the names of the variables to include in the occupancy
process in the occupancy model. Random effects and interactions are specified as in
\code{\link[lme4]{lmer}}. Note that only second order interactions are accepted at the moment
(i.e., interactions of two variables).}

\item{det_mod}{A character vector with the names of the variables to include in the detection
process in the occupancy model. Random effects and interactions are specified as in
\code{\link[lme4]{lmer}}. Note that only second order interactions are accepted at the moment.
(i.e., interactions of two variables).}

\item{fixed_vars}{A character vector with the names of the variables included in the
occupancy model that don't change over time.}

\item{mod_file}{Name of file containing model, with out path to directory.
Directory is specified in \code{mod_dir}. This is typically used for JAGS or Stan
where models are written on an external file.}

\item{server}{Logical. If TRUE the preamble is prepared to run remotely,
otherwise it is prepared to run locally.}

\item{data_dir}{Path to data directory. There are a few inputs to the pipeline
that it doesn't generate itself, such as some environmental layers that are not
on Google Earth Engine. Those would be stored here.}

\item{mod_dir}{Path to directory where models are saved.}

\item{out_dir}{Path to output directory. Pipeline outputs will be stored here,
including intermediate outputs, so most what we need is here.}

\item{package}{A character string with the name of the package that should be used for
fitting occupancy or state-space models.}
}
\value{
A list of parameters that will be passed on to other functions in the
pipeline.
}
\description{
Set basic variables to run the BIRDIE pipeline locally or remotely.
}
\examples{
config <- configPipeline(
    year = 2021,
    dur = 29,
    mod_file = "cwac_ssm_two_season_mean_rev.R",
    package = "jagsUI",
    data_dir = NULL,
    out_dir = NULL,
    server = FALSE
    )
}
