% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createCombinedExportFile}
\alias{createCombinedExportFile}
\title{Create a combined file for export to data mart}
\usage{
createCombinedExportFile(config, type = c("abu", "dst", "indtr_dst"))
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS} and \link{configPreambOccuR}}

\item{type}{A character string with three options: "abu", for abundance estimates files,
"dst" for occupancy estimates files, or "indtr_dst" for distribution indicator files.}
}
\value{
It will create a file in config$out_dir combining all files of the selected
type of all species and years in config$years.
}
\description{
Create a combined file for export to data mart
}
