% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createLog}
\alias{createLog}
\title{Create pipeline event log}
\usage{
createLog(
  config,
  log_file = NULL,
  date_time = NULL,
  species = NULL,
  model = NULL,
  year = NA,
  data = NA,
  fit = NA,
  diagnose = NA,
  summary = NA,
  package = NA,
  notes = "",
  full_log = NULL
)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}} and \code{\link{configPipeline}}.}

\item{log_file}{Optional. A character string with the path to the file that
should be updated. If NULL (default) a new log file will be created.}

\item{date_time}{Optional. A character string with the log date. If NULL (default)
the time given by \code{\link{Sys.time}}, formatted as SAST will be used.}

\item{species}{SAFRING code of the species the log corresponds to.}

\item{model}{Type of model being processed. Either "occ" or "ssm".}

\item{year}{Year the model is fit for.}

\item{data}{Data preparation status. Defaults to NA.}

\item{fit}{model fitting status. Defaults to NA.}

\item{diagnose}{Model diagnostics status. Defaults to NA.}

\item{summary}{Model summary status. Defaults to NA.}

\item{package}{A character string with the name of the package that should be
used to fit models. Currently, only \code{spOccupancy} for occupancy and \code{jagsUI}
for state-space modelling are supported.
maintained.}

\item{notes}{Any additional comments. Defaults to NA.}

\item{full_log}{Optionally, we can pass a full vector log, with all the above
elements. In this case, all other arguments but \code{config} and \code{logfile}
will be ignored.}
}
\value{
A .csv will be saved on the reports directory. The location of this
directory is configured in the configuration object passed as \code{config} at
\code{config$output_dir}.
}
\description{
This function will create a customized log entry for a log file
that records the activity of the pipeline. More precisely, it will record the
time, species and outcome of data preparation, model fitting, diagnostics and
preparation.
}
