% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOccuData.R
\name{createOccuData}
\alias{createOccuData}
\title{Create data for fitting an occupancy model}
\usage{
createOccuData(config, sp_code, years, site_data, visit_data = NULL)
}
\arguments{
\item{config}{A list of configuration parameters see \code{\link{configPipeline}}}

\item{sp_code}{SAFRING_No of the species of interest as extracted from ABAP.
Ignored if download set to FALSE.}

\item{years}{A numeric vector with elements corresponding to the years we
want data for. Ignored if download set to FALSE.}

\item{site_data}{A dataframe with occupancy site data.}

\item{visit_data}{Optional. A dataframe with occupancy visit data.}
}
\value{
A list containing two data frames: one with site data and one with
visit data, if visit data is provided. The second element will be NULL if visit
data not provided. Note that if \code{visit_data} is provided then all the sites without
visits will be removed - i.e., both site and visit data frames will have the same
set of sites.
}
\description{
This function takes generic occupancy site data and, optionally,
visit data without species observations, adds detection data for a species in a given
year, runs some checks and formats the output.
}
