% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOccuRData.R
\name{createOccuRData}
\alias{createOccuRData}
\title{Create data for fitting a model with occuR}
\usage{
createOccuRData(
  sp_code,
  years,
  site_data,
  visit_data,
  force_abap_dwld = FALSE,
  save_occu_data = TRUE,
  overwrite_occu_data = NULL,
  config
)
}
\arguments{
\item{sp_code}{SAFRING_No of the species of interest as extracted from ABAP.
Ignored if download set to FALSE.}

\item{years}{A numeric vector with elements corresponding to the years we
want data for. Ignored if download set to FALSE.}

\item{site_data}{A dataframe with occupancy site data.}

\item{visit_data}{A dataframe with occupancy visit data.}

\item{force_abap_dwld}{Indicates whether ABAP data must be downloaded for the species
and years indicated by 'sp_code' and 'years'. If TRUE, data will be downloaded
from ABAP once per session and cached in a temp file. After this the cached
file will be used, unless download is set to FALSE, in which case data will
be downloaded regardless of the cached file.}

\item{save_occu_data}{If TRUE data is saved to disc, but see 'overwrite_occu_data'.}

\item{overwrite_occu_data}{A character vector with the data that should be overwritten in
case it is already present on disc. It can be any combination of
c("site", "visit", "det"), with site referring to site data, visit to visit data
and det to detection data. Site and visit data are typically common for
multiple species and we might not want to save it all the time.}

\item{config}{A list of configuration parameters see \link{configPreambOccuR}}
}
\value{
A list containing two data frames: one with site data and one with
visit data ready to use with the occuR package
}
\description{
This function takes generic occupancy visit and site data
without species observations, adds detection data for a species in a given
year and formats the output to match occuR standards.
}
