% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredFromAbap.R
\name{createPredFromAbap}
\alias{createPredFromAbap}
\title{Create prediction object directly from ABAP data}
\usage{
createPredFromAbap(sp_code, year_sel, config)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year_sel}{Year in data to run model for.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{fit}{An occupancy model fit to summarise occupancy and detection
probabilities from.}
}
\description{
This function is used when it is not possible to run an occupancy model for
an species and year. Then, the raw data is presented but they must have the
same structure as the prediction object to be able to integrate them in the
database.
}
