% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jags.R
\name{diagnoseGofJagsSsm}
\alias{diagnoseGofJagsSsm}
\title{Diagnose goodness-of-fit for JAGS state-space model}
\usage{
diagnoseGofJagsSsm(fit, counts, linear = TRUE)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{counts}{A dataframe with CWAC counts ready for model fit. See \code{\link{ppl_create_data_ssm}}}

\item{linear}{Logical. If TRUE (default) then posterior predictive checks are
made on the data scale, otherwise they are done in the model (log) scale.}
}
\value{
A data frame with Rhat values for the different parameters estimated
by the model.
}
\description{
This function runs basic posterior predictive checks for a JAGS SSM
}
