% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spOccupancy.R
\name{diagnoseRhatSpOccu}
\alias{diagnoseRhatSpOccu}
\title{Diagnose convergence for spOccupancy model}
\usage{
diagnoseRhatSpOccu(fit, sp_code, year)
}
\arguments{
\item{fit}{A \code{spOccupancy} model fit to ABAP data.}

\item{sp_code}{SAFRING reference number of the species we want to analyse.}

\item{year}{The year the pipeline is run for}
}
\value{
A data frame with Rhat values for the different parameters estimated
by the model.
}
\description{
This function runs basic Rhat checks for an spOccupancy model
}
