% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseSpOccu.R
\name{diagnoseSpOccu}
\alias{diagnoseSpOccu}
\title{Run diagnostics for spOccupancy model}
\usage{
diagnoseSpOccu(fit, sp_code, config, year_sel)
}
\arguments{
\item{fit}{An spOccupancy model fit see \code{\link[spOccupancy]{PGOcc}}}

\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccu}}

\item{year_sel}{The year the model was run for.}
}
\value{
The function will create a list with several objects. The most important are:
\itemize{
\item fit.y: For each MCMC sample, chi-squared statistic for total number of detections
for all sites in data wrt expected model expectation
\item fit.y.rep: For each MCMC sample, chi-squared statistic for total number of detections
for all sites in simulated data wrt expected model expectation
\item fit.y.group.quants: For each site, chi-squared statistic for total number of detections
in data wrt expected model expectation (posterior distribution quantiles)
\item fit.y.rep.group.quants: For each site, chi-squared statistic for total number of detections
in simulated data wrt expected model expectation (posterior distribution quantiles)
\item y.summ.per.site: Data frame with number of detections per site obtained from
data simulated from the posterior distribution. Also the observed number of detections
in the data.
}
}
\description{
This function runs basic diagnostic checks for spOccupancy models
performing posterior predictive goodness of fit checks see \code{\link[spOccupancy]{ppcOcc}}.
}
