% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAoo.R
\name{estimateAoo}
\alias{estimateAoo}
\title{Estimate Area of Occupancy (AOO)}
\usage{
estimateAoo(year, spp_code, site_data, visit_data, config, verbose = TRUE)
}
\arguments{
\item{year}{Year selected to estimate AOO for.}

\item{spp_code}{SAFING_No for the species predicting AOO for, as coming from
the ABAP API.}

\item{site_data}{Site data used to fit the occupancy model from which AOO is
estimated.}

\item{visit_data}{Visit data used to fit the occupancy model from which AOO is
estimated}

\item{config}{A named list coming from \link{configPreambOccuR}.}

\item{verbose}{To silence data update message, set to FALSE.}
}
\value{
The indicator file will be updated with the resulting AOO info, which
includes: spp_code, year, estimate, sd, lower 95\% CI, upper 95\% CI
}
\description{
Estimate Area of Occupancy (AOO)
}
