% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAOOSpType.R
\name{estimateAOOSpType}
\alias{estimateAOOSpType}
\title{Estimate Area Of Occupancy for species type}
\usage{
estimateAOOSpType(sp_type, year, config, force_predict = FALSE, verbose)
}
\arguments{
\item{sp_type}{A character string with the species type for which}

\item{year}{Year for which indicators are to be estimated}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{force_predict}{If TRUE new prediction will be run from models and they
will be stored in a temporary directory. If FALSE cached predictions will be
used if they exist.}

\item{verbose}{Logical. If TRUE, the new line added to the indicator table
is displayed in the console.}
}
\description{
Estimate Area Of Occupancy for species type
}
