% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateOccuSpType.R
\name{estimateOccuSpType}
\alias{estimateOccuSpType}
\title{Estimate occupancy for species type}
\usage{
estimateOccuSpType(sp_type, year, config)
}
\arguments{
\item{sp_type}{A character string with the species type for which}

\item{year}{Year for which indicators are to be estimated}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}
}
\value{

}
\description{
Estimate occupancy for species type
}
