% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{eval_joints}
\alias{eval_joints}
\title{Evaluate the quality of a joint distribution approximation}
\usage{
eval_joints(tgt, metric = c("tv", "tv_col", "tv_row", "mad", "rmse"), ...)
}
\arguments{
\item{tgt}{the true joint distribution}

\item{metric}{how to evaluate the joint distributions: total variation
(\code{tv}), mean absolute deviation (\code{mad}), or root mean-square error
(\code{rmse}).}

\item{...}{named joint distributions.}
}
\value{
a tibble with a row for every argument in \code{...}
}
\description{
Evaluate the quality of a joint distribution approximation
}
