% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactExtractParll.R
\name{exactExtractParll}
\alias{exactExtractParll}
\title{Fast raster extraction in parallel}
\usage{
exactExtractParll(rst, spt, ncores, fun, ...)
}
\arguments{
\item{rst}{A raster or raster stack/brick}

\item{spt}{A spatial polygon object (sf, sfc, SpatialPolygonDataFrame)}

\item{ncores}{Number of cores to use in the extraction}

\item{fun}{Character string naming the function to use for summarizing the
raster values per polygon.}

\item{...}{Other arguments to pass onto the function 'fun'.
exactextractr::exactextract()}
}
\value{
An spatial polygon object with an extract column containing the
summary of raster values per polygon
}
\description{
This function is a wrapper around exactextractr::exactextract()
to perform extraction of raster values within polygons in parallel. It
divides the data into blocks and performs extraction using one core per block.
}
