% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParRcppocc.R
\name{extractParRcppocc}
\alias{extractParRcppocc}
\title{Extract samples from a parameter in Rcppocc multi-chain fit}
\usage{
extractParRcppocc(fit, pars)
}
\arguments{
\item{fit}{Multichain Rcppocc fit obtained via prllOccSPATlogit}

\item{par}{Parameter we want to extract samples from}
}
\value{
A dataframe with the samples of the parameter of interest with a
column indicating the chain the samples correspond to.
}
\description{
Extract samples from a parameter in Rcppocc multi-chain fit
}
