% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVisitCovt.R
\name{extractVisitCovt}
\alias{extractVisitCovt}
\title{Extract raster covariates for occupancy visit data}
\usage{
extractVisitCovt(.sub_by, .visit_data, .sites, .covt, .name, .p = p)
}
\arguments{
\item{.sub_by}{A character string corresponding to the name used to subset
the covariate (e.g. "February", "2015", etc.)}

\item{.visit_data}{A dataframe with occupancy visit data.}

\item{.sites}{A spatial object with the sites sampled to obtain .visit_data.}

\item{.covt}{A raster layer with the covariate to extract. If .covt is a
rasterStack, the layer with name equal to .sub_by with be extracted.}

\item{.name}{Name that will be given to the new extracted covariate}

\item{.p}{This is needed for progress reports. Just leave the default.}
}
\value{
The .visit_data dataframe with an additional column with the new
variable
}
\description{
Extract raster covariates for occupancy visit data
}
