% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNextIndex.R
\name{findNextIndex}
\alias{findNextIndex}
\title{Find index of next record matching criteria}
\usage{
findNextIndex(x, find_id)
}
\arguments{
\item{x}{A vector.}

\item{find_id}{A vector containing the elements that indices want to be
found of.}
}
\value{
A vector with the same length of x containg, for each element in x,
what is the index of the next record matching criteria passed on to find_id.
}
\description{
The function takes a vector and, for each element, finds the
index of the next record that matches certain criteria.
}
\examples{
x <- rbinom(20, 1, 0.2)
findNextIndex(x, 1)
}
