% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCwacSsm.R
\name{fitCwacSsm}
\alias{fitCwacSsm}
\title{Fit a state-space model to CWAC counts}
\usage{
fitCwacSsm(counts, mod_file = NULL, param, jags_control = NULL)
}
\arguments{
\item{counts}{A data frame with at least two columns: "counts" - an integer column corresponding to the counts of a year and season
and "season_id" - an integer column that identifies the season (1 for summer and 2 for winter).}

\item{mod_file}{A character string corresponding to the directory where the JAGS model lives at. If not provided, writeJagsModelFile is called to create a "default" model.}

\item{param}{A vector with names of parameters to monitor.}

\item{jags_control}{A list specifying the JAGS MCMC settings. Possible options are: inits - initial parameter values,
ni - total number of iterations, nb - number of iterations to burn, nt - chain thinning, nc - number of chains,
na - number of adapting iterations, ncores - number of cores to use.}
}
\value{
A JAGS fit object
}
\description{
Fit a state-space model to CWAC counts
}
\examples{
\dontrun{
counts <- barberspan
ssmcounts <- prepSsmData(counts, species = NULL)
fitCwacSsm(ssmcounts, mod_file = "mymodel.jags",
param = c("beta", "lambda", "sig.zeta",
"sig.w", "sig.eps", "sig.alpha", "sig.e", "mu_t", "mu_wt"))
}
}
