% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOccuR.R
\name{fitOccuR}
\alias{fitOccuR}
\title{Fit occuR occupancy model}
\usage{
fitOccuR(site_data_year, visit_data_year, config, spatial = FALSE, sp_sites)
}
\arguments{
\item{site_data_year}{Occupancy site data for a single year and species
(see \code{\link{ppl_create_site_visit}})}

\item{visit_data_year}{Occupancy visit data for a single year and species
(see \code{\link{ppl_create_site_visit}})}

\item{config}{A list with pipeline configuration parameters
(see \code{\link{configPreambOccu}}).}

\item{spatial}{Logical, indicating whether spatial random effects should be
included in the model (TRUE) or not (FALSE, default)}

\item{sp_sites}{Spatial object containing the pentads in \code{site_data_year}.}
}
\value{
Either an occuR model fit or the integer 3, indicating that model fit
failed.
}
\description{
Fit occuR occupancy model
}
