% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spOccupancy.R
\name{fitSpOccu}
\alias{fitSpOccu}
\title{Fit spOccupancy model}
\usage{
fitSpOccu(
  site_data_year,
  visit_data_year,
  config,
  sp_code,
  spatial = FALSE,
  sp_sites,
  ...
)
}
\arguments{
\item{site_data_year}{Occupancy site data for a single year and species
(see \code{\link{ppl_create_site_visit}})}

\item{visit_data_year}{Occupancy visit data for a single year and species
(see \code{\link{ppl_create_site_visit}})}

\item{config}{A list with pipeline configuration parameters
(see \code{\link{configPreambOccu}}).}

\item{sp_code}{SAFRING code of the species the pipeline is running for}

\item{spatial}{Logical, indicating whether spatial random effects should be
included in the model (TRUE) or not (FALSE, default)}

\item{sp_sites}{Spatial object containing the pentads in \code{site_data_year}.}

\item{...}{Other arguments that might be needed (e.g. for messages)}
}
\value{
Either a spOccupancy model fit or the integer 3, indicating that model fit
failed.
}
\description{
Fit spOccupancy model
}
