% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherYearFromVars.R
\name{gatherYearFromVars}
\alias{gatherYearFromVars}
\title{Gather years from variables with year}
\usage{
gatherYearFromVars(x, vars, sep)
}
\arguments{
\item{x}{A dataframe with columns that represent a variable on a given year.
Note that there must be a unique key that identifies each case uniquely.}

\item{vars}{Variables to separate in variable and year (in different columns).}

\item{sep}{A character that identifies the separator between the name of the
variable and the year.}
}
\value{
A dataframe with separate columns for variables and years. There will
be a column that represents the year and several columns that represent the
variables measured across years.
}
\description{
Gather years from variables with year
}
\examples{
df <- data.frame(id = 1:10,
                 v1_2010 = rnorm(10, 0, 1),
                 v1_2011 = rnorm(10, 100, 10),
                 v2_2010 = runif(10, 0, 1),
                 v2_2011 = runif(10, 100, 110))

gatherYearFromVars(x = df, vars = names(df)[-1], sep = "_")
}
