% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOccVisitData.R
\name{getOccVisitData}
\alias{getOccVisitData}
\title{Get detection/non-detection data}
\usage{
getOccVisitData(region_type, region, species, path)
}
\arguments{
\item{region_type}{The type of region we are interested on.
Three options: "country", "province" and "pentad".}

\item{region}{A character string corresponding to the specific region we are
interested in. It can be either a country in Southern Africa, a South African
province or a pentad code.}

\item{species}{The code of the species we are interested in.}

\item{path}{A directory where administrative boundaries layer should be
looked for.}
}
\value{

}
\description{
This is a wrapper around getRegionPentads and
SABAP::getSabapData that prepares SABAP detection/non-detection data to be
analysed in an occupancy model.
}
