% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSpCoastal.R
\name{isSpCoastal}
\alias{isSpCoastal}
\title{Is species coastal}
\usage{
isSpCoastal(sp_sel, out_dir, visit_mod)
}
\arguments{
\item{sp_sel}{The SAFRING code of the species being assessed.}

\item{out_dir}{The output directory where information must be saved and where
the diagnostic file should be found.}

\item{visit_mod}{The model for the detection probability. The occupancy model
is fixed to \code{psi ~ 1 + dist_coast}.}

\item{force}{Logical; if FALSE the function will check if the an assessment
has been done for this species already and skip if it has. See details.}
}
\value{
It returns TRUE if the species is coastal and it will create a file
in \code{out_dir} to indicate that the check has already been done for this
species. In the future, the function will check if this file exits before
and avoid fitting the model if it does, unless force == TRUE
}
\description{
This function is used to determine whether a species is coastal.
It fits a simple occupancy model with distance to coast as a covariate for
occupancy probability. If the coefficient is large (>2) then the species is
considered coastal and a diagnostic file is created in a directory selected
by the user.
}
\examples{
\dontrun{
isSpCoastal(235, outdir, reformulate(c("1", "log(TotalHours+1)"), response = "p"))
}
}
