% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logFitStatus}
\alias{logFitStatus}
\title{Generate fit status messages}
\usage{
logFitStatus(.fit_out, .year, .sp_code, .config)
}
\arguments{
\item{.fit_out}{Fit status object generated by \code{\link{ppl_fit_occu_model}}}

\item{.year}{Year to run to the pipeline for}

\item{.sp_code}{Species SAFRING code}

\item{.config}{Config object from \code{\link{configPipeline}}}
}
\description{
Generate fit status messages
}
