% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_race.R
\name{model_race}
\alias{model_race}
\title{Estimate X versus Race Given Name, Location, Outcome, and Covariates}
\usage{
model_race(
  r_probs,
  X,
  G,
  Z = NULL,
  condition = NULL,
  data = NULL,
  prefix = "pr_",
  method = c("svi", "mle"),
  config = list(),
  silent = FALSE,
  reload_py = FALSE
)
}
\arguments{
\item{r_probs}{a matrix data frame of race probabilities}

\item{X}{the column containing the outcome}

\item{G}{the column containing locations}

\item{Z}{the column(s), if any, containing other covariates. Use \code{c()} to provide multiple columns.}

\item{condition}{the columns in \code{G} or \code{Z} to condition on in creating predictions.
For example, if \code{Z} contains \code{gender}, setting \code{condition=gender} would
produce X|R estimates for each gender.}

\item{data}{the data}

\item{prefix}{how to select the race probability columns from \code{r_probs}, if
if is a data frame}

\item{config}{a list containing extra arguments to the Pyro model}

\item{silent}{whether to suppress output}

\item{reload_py}{whether to reload the python code (development only)}
}
\value{
A list containing the model output. Element \code{p_xr} contains the
approximate posterior draws of the global X|R table.
}
\description{
Bayesian model for X|R given name and covariates
}
