% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_race_prep.R
\name{model_race_load}
\alias{model_race_load}
\title{Load a fitted model from Python}
\usage{
model_race_load(path, fit_obj = NULL)
}
\arguments{
\item{path}{the path to the directory containing model input/output}

\item{fit_obj}{the output of \link{model_race_prep}}
}
\value{
A {.cls fit_raceproxy} object.
}
\description{
Requires the \code{reticulate} package.
}
