% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_race_prep.R
\name{model_race_prep}
\alias{model_race_prep}
\title{Prepare estimation X versus Race Given Name, Location, Outcome, and Covariates}
\usage{
model_race_prep(
  r_probs,
  X,
  G,
  Z = NULL,
  condition = NULL,
  data = NULL,
  prefix = "pr_",
  method = c("svi", "mle"),
  config = list()
)
}
\arguments{
\item{r_probs}{a matrix data frame of race probabilities}

\item{X}{the column containing the outcome}

\item{G}{the column containing locations}

\item{Z}{the column(s), if any, containing other covariates. Use \code{c()} to provide multiple columns.}

\item{condition}{the columns in \code{G} or \code{Z} to condition on in creating predictions.
For example, if \code{Z} contains \code{gender}, setting \code{condition=gender} would
produce X|R estimates for each gender.}

\item{data}{the data}

\item{prefix}{how to select the race probability columns from \code{r_probs}, if
if is a data frame}

\item{config}{a list containing extra arguments to the Pyro model}
}
\value{
A list containing the prepared data. A sample call will also be printed.
}
\description{
Prepares and writes data for use in Python models.
}
