% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_race_prep.R
\name{model_race_python}
\alias{model_race_python}
\title{Estimate X versus Race Given Name, Location, Outcome, and Covariates}
\usage{
model_race_python(
  r_probs,
  X,
  G,
  Z = NULL,
  condition = NULL,
  data = NULL,
  prefix = "pr_",
  method = c("svi", "mle"),
  path = NULL,
  python_exec = "python3",
  prep_only = FALSE,
  config = list()
)
}
\arguments{
\item{r_probs}{a matrix data frame of race probabilities}

\item{X}{the column containing the outcome}

\item{G}{the column containing locations}

\item{Z}{the column(s), if any, containing other covariates. Use \code{c()} to provide multiple columns.}

\item{condition}{the columns in \code{G} or \code{Z} to condition on in creating predictions.
For example, if \code{Z} contains \code{gender}, setting \code{condition=gender} would
produce X|R estimates for each gender.}

\item{data}{the data}

\item{prefix}{how to select the race probability columns from \code{r_probs}, if
if is a data frame}

\item{method}{whether to use stochastic variational inference ("svi") or a
Laplace approximation ("mle") for inference.}

\item{path}{a directory to save the model outputs to. Ignored if NULL}

\item{python_exec}{name of Python executable}

\item{prep_only}{if \code{TRUE}, don't call Python to fit the model}

\item{config}{a list containing extra arguments to the Pyro model}
}
\value{
if \code{prep_only=FALSE}, the fitted model. Otherwise, a list containing
the prepared data. A sample call will also be printed.
}
\description{
Prepares data for use in Python models and then calls Python through the
system interface.
}
