% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipe_prll_fit}
\alias{pipe_prll_fit}
\title{Fit occupancy models in parallel}
\usage{
pipe_prll_fit(
  .sp_code,
  .year,
  .spatial = FALSE,
  .config,
  .steps = c("fit", "diagnose"),
  time_limit = NULL
)
}
\arguments{
\item{.sp_code}{Species SAFRING code}

\item{.year}{Year to run to the pipeline for}

\item{.spatial}{Whether a spatial model should be fit. Defaults to FALSE.}

\item{.config}{Config object from \code{\link{configPipeline}}}

\item{.steps}{Character vector with the steps the pipeline should run. Typically
c("fit", "diagnose"). However, be aware that "diagnose" uses a lot of RAM.}

\item{time_limit}{Computation time limit for each core in seconds.}
}
\description{
Wrapper around ppl_run_pipe_dst1 (fit step) for furrr parallel calls
}
