% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCtSsm.R
\name{plotCtSsm}
\alias{plotCtSsm}
\title{Plot state series of a continuous-time CWAC state-space model
(EXPERIMENTAL)}
\usage{
plotCtSsm(fit, ssm_counts, dyn = FALSE)
}
\arguments{
\item{fit}{A continuous-time JAGS state-space model fitted to CWAC data}

\item{ssm_counts}{A data frame with the count data use to fit the
state-space model}

\item{dyn}{Whether the fitted long-term trend is fixed or dynamic
(NOT USED AT PRESENT).}
}
\value{
A list with two elements: i) plot: a plot with summer and winter
fitted states, as well as the long-term trend, ii) data: the data used to
create the individual plots (e.g. for exporting to the dashboard).
}
\description{
Plot state series of a continuous-time CWAC state-space model
(EXPERIMENTAL)
}
