% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOccuVarEffect.R
\name{plotOccuVarEffect}
\alias{plotOccuVarEffect}
\title{Plot effect of occupancy covariates}
\usage{
plotOccuVarEffect(fit, new_occu_data, var, type = "occu", nboot = 1000)
}
\arguments{
\item{fit}{An occuR model fit.}

\item{new_occu_data}{An list with occupancy site and visit data to predict
from. See \code{\link[BIRDIE]{prepDataOccuR}}}

\item{var}{Character string with the name of the variables to plot. It
must match those names in new_occu_data.}

\item{type}{A character string of either "occu" if the variable is plotted
against occupancy probabilities or "detc" if the variable is plotted against
detection probabilities}

\item{nboot}{Number of bootstrap samples to compute confidence intervals.}
}
\value{
A plot of the effect of selected variables on occupancy or detection.
probabilities.
}
\description{
Plots the effect of selected covariates
}
