% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOccuVars.R
\name{plotOccuVars}
\alias{plotOccuVars}
\title{Plot occupancy covariates}
\usage{
plotOccuVars(occu_data, vars, type = "ind")
}
\arguments{
\item{occu_data}{An list with occupancy site and visit data. See \code{\link[BIRDIE]{prepDataOccuR}}}

\item{vars}{Character vector with the names of the variables to plot. They
must match those names in occu_data.}

\item{type}{The type of plot we want to display. Either "ind" which shows
individual plots for each variable, or "cross", which shows a scatterplot
with two variables. We need to pass on exactly two variables.}
}
\value{
A plot with detection/non-detection against selected variables
}
\description{
Plots selected covariates against detection/non-detection data
}
