% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSsm.R
\name{plotSsm}
\alias{plotSsm}
\title{Plot state series of a CWAC state-space model}
\usage{
plotSsm(fit, ssm_counts, dyn = FALSE)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{ssm_counts}{A data frame with the count data use to fit the state-space model}

\item{dyn}{Whether the fitted long-term trend is fixed or dynamic (NOT USED AT PRESENT).}
}
\value{
A plot with summer and winter fitted states, as well as the long-term trend
}
\description{
Plot state series of a CWAC state-space model
}
