% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSsm2ss.R
\name{plotSsm2ss}
\alias{plotSsm2ss}
\title{Plot state series of a CWAC state-space model with two seasons}
\usage{
plotSsm2ss(fit, ssm_counts, dyn = FALSE)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{ssm_counts}{A data frame with the count data use to fit the state-space model}

\item{dyn}{Whether the fitted long-term trend is fixed or dynamic (NOT USED AT PRESENT).}
}
\value{
A plot with summer and winter fitted states, as well as the long-term trend
}
\description{
Plot state series of a CWAC state-space model with two seasons
}
\examples{
counts <- barberspan
ssmcounts <- prepSsmData(counts, species = NULL)
fit_fxd <- fitCwacSsm2ss(ssmcounts,
                         mod_file = "analysis/models/cwac_ssm_2ss_fxd.jags",
                         param = c("beta", "sig.w", "sig.eps", "sig.alpha",
                                   "sig.e", "mu_t", "mu_wt"))
plotSsm(fit = fit_fxd, ssm_counts = ssmcounts)

}
