% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jags.R
\name{postPredDistJagsSsm}
\alias{postPredDistJagsSsm}
\title{Prepare posterior predictive distribution for JAGS state-space model}
\usage{
postPredDistJagsSsm(fit, data, nsamples)
}
\arguments{
\item{fit}{A state-space model fitted with JAGS}

\item{data}{The data used to fit the model \code{fit}. This must be counts from
CWAC data. See \code{\link{ppl_fit_ssm_model}}}

\item{nsamples}{Number of posterior samples that should be used to build the
data frame. Defaults to 500.}
}
\value{
A data frame with the observed response and the corresponding estimates
obtained from a Bayesian state-space model fitted with JAGS. The variable \code{obs_sim}
correspond contains the real and simulated data. The variable \code{iter} identifies the
iteration each observation corresponds to, and those observation with \code{iter = 0}
and/or \code{obs = 1} correspond to observed data.
}
\description{
This function prepares a data frame with the observed response
and the corresponding estimates obtained from a Bayesian state-space model
fitted with JAGS
}
