% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_create_data_ssm.R
\name{ppl_create_data_ssm}
\alias{ppl_create_data_ssm}
\title{Prepare covariate data for abundance pipeline}
\usage{
ppl_create_data_ssm(sp_code, year, catchment, config, ...)
}
\arguments{
\item{year}{Year to run to the pipeline for}

\item{catchment}{A sf object with polygons corresponding to the catchment, or
reference area considered for the covariates associated with CWAC sites.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{...}{Other arguments passed on to \link{prepGEESpCountData}}
}
\value{

}
\description{
Prepare covariate data for abundance pipeline
}
