% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_create_indtr_files.R
\name{ppl_create_indtr_files}
\alias{ppl_create_indtr_files}
\title{Create indicator storage files}
\usage{
ppl_create_indtr_files(config, overwrite_indtr, ...)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{overwrite_indtr}{Logical. If TRUE, existing files in directories
corresponding to the species in config$species will be overwritten.}

\item{...}{}
}
\value{

}
\description{
Create indicator storage files
}
