% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_create_site_visit.R
\name{ppl_create_site_visit}
\alias{ppl_create_site_visit}
\title{Create site and visit occupancy files}
\usage{
ppl_create_site_visit(sp_code, year, config, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Create site and visit occupancy files
}
