% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_data_ssm.R
\name{ppl_data_ssm}
\alias{ppl_data_ssm}
\title{Prepare covariate data for abundance pipeline}
\usage{
ppl_data_ssm(site, year, config, ...)
}
\arguments{
\item{site}{Code for site of interest.}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Prepare covariate data for abundance pipeline
}
