% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_diagnose_occu.R
\name{ppl_diagnose_occu}
\alias{ppl_diagnose_occu}
\title{Run diagnostics for occupancy model fit}
\usage{
ppl_diagnose_occu(fit, data = NULL, sp_code, year)
}
\arguments{
\item{fit}{A \code{spOccupancy} model fit to ABAP data.}

\item{data}{A dataset with detection/non-detection data. Not used at the moment because
spOccupancy (currently used) fit objects contain the data used to fit the model.}

\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year}{The year the pipeline is run for}
}
\value{
A data frame with Rhat values for the different parameters estimated
by the model is returned.
}
\description{
This function performs basic checks for an occupancy model fit,
such as convergence of parameters and posterior predictive checks
}
