% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_diagnose_ssm.R
\name{ppl_diagnose_ssm}
\alias{ppl_diagnose_ssm}
\title{Run diagnostics for JAGS state-space model fit}
\usage{
ppl_diagnose_ssm(fit, counts, sp_code, config)
}
\arguments{
\item{sp_code}{SAFRING reference number of the species we want to analyze.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}
}
\value{
A data frame with Rhat values for the different parameters estimated
by the model, will be saved to disk.
}
\description{
This function performs basic checks for a JAGS SSM fit, such as
convergence of parameters and posterior predictive checks
}
