% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_estimate_daoo.R
\name{ppl_estimate_daoo}
\alias{ppl_estimate_daoo}
\title{Estimate change in Area of Occupancy}
\usage{
ppl_estimate_daoo(
  sp_code,
  config,
  term = c("annual", "short", "long"),
  verbose,
  ...
)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{term}{A character string with one of three options with the time-frame
for the change: 'annual' (1 year), 'short' (15 years), 'long' (30 years).}

\item{verbose}{Logical. If TRUE, the new line added to the indicator table
is displayed in the console.}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Estimate change in Area of Occupancy
}
