% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_estimate_distr_sp_type.R
\name{ppl_estimate_distr_sp_type}
\alias{ppl_estimate_distr_sp_type}
\title{Estimate distribution indicators for species type}
\usage{
ppl_estimate_distr_sp_type(
  sp_type,
  year,
  config,
  force_predict = FALSE,
  verbose,
  ...
)
}
\arguments{
\item{sp_type}{A character string with the species type for which}

\item{year}{Year for which indicators are to be estimated}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{force_predict}{If TRUE new prediction will be run from models and they
will be stored in a temporary directory. If FALSE cached predictions will be
used if they exist.}

\item{verbose}{Logical. If TRUE, the new line added to the indicator table
is displayed in the console.}

\item{...}{}
}
\description{
Estimate distribution indicators for species type
}
