% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_fit_ssm_model.R
\name{ppl_fit_ssm_model}
\alias{ppl_fit_ssm_model}
\title{Fit state-space JAGS model}
\usage{
ppl_fit_ssm_model(sp_code, config, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species of interest}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{...}{Other arguments passed on to other functions}
}
\description{
Fit state-space JAGS model
}
