% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_prep_occur_data.R
\name{ppl_prep_occur_data}
\alias{ppl_prep_occur_data}
\title{Prepare occuR data for model fitting}
\usage{
ppl_prep_occur_data(sp_code, year, config, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Prepare occuR data for model fitting
}
