% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_run_pipe_abu.R
\name{ppl_run_pipe_abu}
\alias{ppl_run_pipe_abu}
\title{Run abundance indicators pipeline}
\usage{
ppl_run_pipe_abu(sp_code, year, site, config, steps = c("fit", "summ"), ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year}{Year to run to the pipeline for}

\item{site}{Code for site of interest.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{steps}{A character vector containing the steps of the pipeline to run.
Can contain: "fit", "summ". Defaults to all of them.}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Run abundance indicators pipeline
}
