% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_run_pipe_abu2.R
\name{ppl_run_pipe_abu2}
\alias{ppl_run_pipe_abu2}
\title{Run abundance indicators pipeline for a site}
\usage{
ppl_run_pipe_abu2(site, year, config, steps = c("data", "fit", "summ"), ...)
}
\arguments{
\item{site}{Code for site of interest.}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{steps}{A character vector containing the steps of the pipeline to run.
Can contain: "data", "fit", "summ". Defaults to all of them.}

\item{...}{Other arguments passed on to other functions}
}
\description{
Run abundance indicators pipeline for a site
}
