% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_run_pipe_dst1.R
\name{ppl_run_pipe_dst1}
\alias{ppl_run_pipe_dst1}
\title{Run distribution indicators pipeline module 1}
\usage{
ppl_run_pipe_dst1(
  sp_code,
  year,
  config,
  steps = c("data", "fit", "diagnose", "summary"),
  ...
)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{steps}{A character vector containing the steps of the pipeline to run.
Can contain: "data", "fit", "diagnose", "summary". Defaults to all of them.}

\item{...}{Other arguments passed on to other functions}
}
\description{
Module 1 of the distribution indicators pipeline estimates
occupancy probabilities in South Africa for a selected species. These
occupancy probabilities form the basis for building more elaborated
indicators in other pipeline modules.
}
