% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_run_pipe_dst2.R
\name{ppl_run_pipe_dst2}
\alias{ppl_run_pipe_dst2}
\title{Run distribution indicators pipeline branch 2}
\usage{
ppl_run_pipe_dst2(
  sp_code,
  indtr = c("aoo", "daoo"),
  config,
  overwrite_indtr,
  verbose,
  ...
)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{indtr}{A character vector containing the indicators of the pipeline
to compute. Can contain: "aoo", "daoo". Defaults to all of them.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{overwrite_indtr}{Logical. If TRUE, existing files in directories
corresponding to the species in config$species will be overwritten.}

\item{verbose}{Logical. If TRUE, the new line added to the indicator table
is displayed in the console.}
}
\description{
Run distribution indicators pipeline branch 2
}
