% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_summarise_occu.R
\name{ppl_summarise_occu}
\alias{ppl_summarise_occu}
\title{Summarise predictions from occupancy model}
\usage{
ppl_summarise_occu(fit, sp_code, year_sel, config, ...)
}
\arguments{
\item{fit}{An occupancy model fit to summarise occupancy and detection
probabilities from.}

\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year_sel}{Year in data to run model for.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{...}{Other arguments passed on to other functions}
}
\description{
Summarise predictions from occupancy model
}
