% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_summarise_ssm.R
\name{ppl_summarise_ssm}
\alias{ppl_summarise_ssm}
\title{Summarise predictions from state-space JAGS model}
\usage{
ppl_summarise_ssm(fit, counts, sp_code, linear = TRUE, config, ...)
}
\arguments{
\item{fit}{A model fit generated by \code{\link{ppl_fit_ssm_model}}}

\item{counts}{A data frame with counts generated by \code{\link{ppl_create_data_ssm}}}

\item{sp_code}{SAFRING code of the species of interest}

\item{linear}{If TRUE (default) then summaries are given the linear scale.
If FALSE, then summaries are given in the modelling scale (typically log scale)}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{...}{Other parameters to pass on to \code{\link{prepGEECatchmData}}}
}
\description{
Summarise predictions from state-space JAGS model
}
