% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_summarize_occur.R
\name{ppl_summarize_occur}
\alias{ppl_summarize_occur}
\title{Summarize prediction from occuR model}
\usage{
ppl_summarize_occur(sp_code, sp_name, year, config, ...)
}
\description{
Summarize prediction from occuR model
}
