% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{predict_race_sgz}
\alias{predict_race_sgz}
\title{Infer Race Given Name, Location, Outcome, and Covariates}
\usage{
predict_race_sgz(
  S,
  G,
  Z = NULL,
  data = NULL,
  p_rs = NULL,
  p_rgz = NULL,
  p_r = c(white = 0.63, black = 0.121, hisp = 0.173, asian = 0.0478, aian = 0.0072, other
    = 0.021),
  est_r_gz = TRUE,
  iterate = 8L * (!est_r_gz & nrow(data) > 100),
  return_gzr = TRUE
)
}
\arguments{
\item{S}{the column containing surnames}

\item{G}{the column containing ZIP codes}

\item{Z}{the column(s), if any, containing other covariates. Use \code{c()} to provide multiple columns}

\item{data}{the data}

\item{p_rs}{a data frame, containing a column matching \code{S} and columns for
each value of \code{R} giving the conditional probabilities of R given S.
Defaults to a table from the U.S. Census Bureau from 2010.}

\item{p_rgz}{a data frame, containing columns matching \code{G}and \code{Z}, and columns for
each value of \code{R} giving the conditional probabilities of R given Z and G.
Defaults to a ZIP code table from the 2010 decennial census.}

\item{p_r}{a vector containing the marginal probabilities for each value of
\code{R}. Defaults to the demographics of the US if \code{est_r_gz} is \code{FALSE}.}

\item{est_r_gz}{If \code{TRUE} estimate race prior \code{p_r} from distribution of \code{G} and \code{Z} in sample}

\item{iterate}{how much to iteratively refine the race prior from the data. Set to 0 to disable.}

\item{return_gzr}{whether to return an estimated p(G,Z|R) matrix as the
\code{p_gzr} attribute and combined G,Z vectors as the \code{gz} attribute.
Set to \code{FALSE} to save space if these values are not needed.
They are required to use \link{calc_joint_bisgz_ols}}
}
\value{
a tibble, with rows matching \code{data} and columns for the race probabilities
}
\description{
BISG(Z) race probability calculations.
}
