% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meas_error.R
\name{predict_race_sgz_me}
\alias{predict_race_sgz_me}
\title{Infer Race Given Name, Location, Outcome, and Covariates}
\usage{
predict_race_sgz_me(
  S,
  G,
  Z = NULL,
  data = NULL,
  p_rs = NULL,
  p_rgz = NULL,
  p_r = c(white = 0.63, black = 0.121, hisp = 0.173, asian = 0.0478, aian = 0.0072, other
    = 0.021),
  iter = 2000,
  warmup = 100
)
}
\arguments{
\item{S}{the column containing surnames}

\item{G}{the column containing ZIP codes}

\item{Z}{the column(s), if any, containing other covariates. Use \code{c()} to provide multiple columns}

\item{data}{the data}

\item{p_rs}{a data frame, containing a column matching \code{S} and columns for
each value of \code{R} giving the number of voters in each level of R given S.
Defaults to a table from the U.S. Census Bureau from 2010.}

\item{p_rgz}{a data frame, containing columns matching \code{G}and \code{Z}, and columns for
each value of \code{R} giving the number of voters in each level of R given Z and G.
Defaults to a table from the 2010 decennial census.}

\item{p_r}{a vector containing the marginal probabilities for each value of
\code{R}. Defaults to the demographics of the US.}

\item{iter}{how many Gibbs iterations}

\item{warmup}{how many warmup iterations}
}
\value{
a tibble, with rows matching \code{data} and columns for the race probabilities
}
\description{
BISG(Z) race probability calculations using a measurement error model.
}
