% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-occuR.R
\name{predictOccuR}
\alias{predictOccuR}
\title{Predict from occuR model fit}
\usage{
predictOccuR(fit, sp_code, year, config, ...)
}
\arguments{
\item{fit}{An occupancy model fit to summarise occupancy and detection
probabilities from.}

\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccu}}

\item{...}{Other arguments passed on to other functions}
}
\value{
A list with two elements: 1) posterior occupancy probability samples for
the South African pentads, and 2) posterior detection probability samples for each
visit in the ABAP data for the year \code{year_sel}.
}
\description{
Predict from occuR model fit
}
