% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spOccupancy.R
\name{predictSpOccu}
\alias{predictSpOccu}
\title{Predict from spOccupancy model fit}
\usage{
predictSpOccu(fit, sp_code, year_sel, config, ...)
}
\arguments{
\item{fit}{An occupancy model fit to summarise occupancy and detection
probabilities from.}

\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year_sel}{Year in data to run model for.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{...}{Other arguments passed on to other functions}
}
\value{
A list with two elements: 1) posterior occupancy probability samples for
the South African pentads, and 2) posterior detection probability samples for each
visit in the ABAP data for the year \code{year_sel}.
}
\description{
Predict from spOccupancy model fit
}
