% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCtSsmData.R
\name{prepCtSsmData}
\alias{prepCtSsmData}
\title{Prepare CWAC data to fit a continuous-time state-space model}
\usage{
prepCtSsmData(counts, species = NULL)
}
\arguments{
\item{counts}{A data frame with raw CWAC count data}

\item{species}{An optional vector of species codes. Only records for these
species are prepared for fitting an SSM, others are discarded.}
}
\value{
A tibble with clean and prepared data for fitting a continuous-time
SSM (e.g. filled with missing years)
}
\description{
Prepare CWAC data to fit a continuous-time state-space model
}
\examples{
counts <- barberspan
prepCtSsmData(counts)
prepCtSsmData(counts, species = 212)
prepCtSsmData(counts, species = c(212, 50))
}
