% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDataOccuR.R
\name{prepDataOccuR}
\alias{prepDataOccuR}
\title{Prepare data for fitting a model with occuR}
\usage{
prepDataOccuR(site_data, visit_data)
}
\arguments{
\item{site_data}{A dataframe with occupancy site data.}

\item{visit_data}{A dataframe with occupancy visit data.}
}
\value{
A list containing two data frames: one with site data and one with
visit data ready to use with the occuR package
}
\description{
Prepare data for fitting a model with occuR
}
