% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGEECatchmData.R
\name{prepGEECatchmData}
\alias{prepGEECatchmData}
\title{Prepare Google Earth Engine data for catchments}
\usage{
prepGEECatchmData(
  sp_code,
  catchment,
  config,
  monitor = TRUE,
  upload_catchment = FALSE,
  force_gee = FALSE
)
}
\arguments{
\item{sp_code}{SAFRING reference number of the species we want to analyze.}

\item{catchment}{An sf object with the polygons defining the catchments
to be annotated.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{monitor}{Logical. If TRUE (default) monitoring printed messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}

\item{upload_catchment}{If TRUE catchment will be uploaded to GEE. FALSE is
an option because catchment info might have been already uploaded to GEE.
Defaults to FALSE.}

\item{force_gee}{If TRUE annotation with GEE info will occur even if a
dataset with covariates is already present on disk.}
}
\description{
Prepare Google Earth Engine data for catchments
}
