% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepOccuData.R
\name{prepOccuData}
\alias{prepOccuData}
\title{Prepare Spatial Occupancy Data}
\usage{
prepOccuData(.occudata, .year, .covts)
}
\arguments{
\item{.occudata}{A dataframe with at least seven columns: "Pentad", "lon",
"lat", "StartDate", "TotalHours", "detc" (1 for detection, 0 for
non-detection).}

\item{.year}{The year that we want to analyse}

\item{.covts}{A data frame with as many rows as sites (Pentads) and
site-level covariates that we want to use for fitting the model. The
covariates must have the year they correspond to in their name.}
}
\value{
An object of class so.data that we can feed into the function
stocc::spatial.occupancy.
}
\description{
Prepares data to fit a spatial occupancy model using the stocc
package.
}
