% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepOccVisitData.R
\name{prepOccVisitData}
\alias{prepOccVisitData}
\title{Prepare occupancy visit data}
\usage{
prepOccVisitData(
  region = NULL,
  sites,
  species,
  years,
  clim_covts,
  covts_dir,
  file_fix,
  downld_path = NULL,
  savedir = NULL
)
}
\arguments{
\item{region}{A character string defining the region of interest. At the
moment only countries are supported. SABAP pentads contained in the region
will be used as sites.}

\item{sites}{An sf object with the sites we are interested in. If provided,
'region' is ignored.}

\item{species}{The SABAP numeric code of the species of interest.}

\item{years}{A numeric vector of years we want to estimate occupancy for.
Covariates for these years will be appended to the data.}

\item{clim_covts}{A character vector with the abbreviation of the climatic
covariates we want to append to the data. At the moment, the Terraclimate
data set is used. See \link[climateR]{getTerraClim} for the variables
available.}

\item{covts_dir}{Directory where covariate data are found.}

\item{file_fix}{A character vector with two elements corresponding to
additional characters that are found before and after clim_covts.}

\item{downld_path}{If provided administrative boundaries layer will be looked
for and stored at here.}

\item{savedir}{Directory where sites should be saved at before covariate
extraction. Intersecting region with SABAP pentads is time consuming, so you
may want to save the selected pentads as soon as possible. The directory
must include the name and extension of the file (.rds).}
}
\value{

}
\description{
Prepare occupancy visit data
}
