% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepPredictDataOccuR.R
\name{prepPredictDataOccuR}
\alias{prepPredictDataOccuR}
\title{Prepare site data to predict from occuR occupancy model}
\usage{
prepPredictDataOccuR(occuRdata, pred_sites, years, scaling = FALSE)
}
\arguments{
\item{occuRdata}{A list produced by \link{prepDataOccuR}}

\item{pred_sites}{A dataframe with the sites to predict occupancy for.}

\item{years}{A vector with years to predict for.}

\item{scaling}{Logical; whether it is necessary to scale the variables
before predicting. Scaling factors will be extracted from the attributes of
occuRdata. See \link{scale}.}
}
\value{
A dataframe with site information to predict from an occuR model.
Sites and occasions are formatted according to occuRdata.
}
\description{
Prepare site data to predict from occuR occupancy model
}
