% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSsmData.R
\name{prepSsmData}
\alias{prepSsmData}
\title{Prepare CWAC data to fit a state-space model}
\usage{
prepSsmData(counts)
}
\arguments{
\item{counts}{A data frame with raw CWAC count data}
}
\value{
A tibble with clean and prepared data for fitting an SSM (e.g. filled with missing years)
}
\description{
Prepare CWAC data to fit a state-space model
}
\examples{
counts <- getCwacSiteCounts(26352535)
prepSsmData(counts)
}
