% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prllOccSPATlogit.R
\name{prllOccSPATlogit}
\alias{prllOccSPATlogit}
\title{Parallel Spatial Occupancy with Rcppocc}
\usage{
prllOccSPATlogit(nchains, ...)
}
\arguments{
\item{nchains}{Number of chains to run.}

\item{...}{Arguments to be used by Rcppocc::occSPATlogit}
}
\value{
A list with nchains outputs from Rcppocc
}
\description{
Wrapper around Rcppocc::occSPATlogit to run multiple chains.
They may run in parallel if future::plan is configured to do so previously.
}
