% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jags.R
\name{processJAGSoutput}
\alias{processJAGSoutput}
\title{Process JAGS model outputs}
\usage{
processJAGSoutput(fit, DIC, params.omit, verbose = TRUE)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{DIC}{Logical stating whether DIC should be computed}

\item{params.omit}{A character vector with the name of the parameters
that should not be processed.}

\item{verbose}{Logical. If TRUE (default), then several messages are displayed
during processing.}
}
\value{
A list with two elements: i) plot: a plot with summer and winter fitted states, as well as the long-term trend,
ii) data: the data used to create the individual plots. This is useful for extracting the data used by ggplot to render the plots
(e.g. for exporting to the dashboard)
}
\description{
Process JAGS model outputs
}
