% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleOccuRVars.R
\name{scaleOccuRVars}
\alias{scaleOccuRVars}
\title{Scale variables in occuR-like data}
\usage{
scaleOccuRVars(site_data, visit_data, scale_vars)
}
\arguments{
\item{site_data}{A dataframe with occupancy site data.}

\item{visit_data}{A dataframe with occupancy visit data.}

\item{scale_vars}{A named list with names 'site' and 'visit'. Each element
must contain the variables to be scaled in the site and visit data
respectively. If one element is NULL, no variables will be scaled in the data
corresponding to that element.}
}
\value{
A named list containing scaled and centred 'site' and 'visit' data.
}
\description{
Scale variables in occuR-like data
}
