% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSppFromDiag.R
\name{selectSppFromDiag}
\alias{selectSppFromDiag}
\title{Select species based on model diagnostics}
\usage{
selectSppFromDiag(config, sp_codes, year, module = c("dst", "abu"))
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPipeline}.}

\item{sp_codes}{SAFRING reference numbers of the species we want diagnostics for.}

\item{year}{The year for which diagnostics are required.}

\item{module}{Either "dst" if we want occupancy diagnostics or "abu" if we want
state-space model diagnostics.}
}
\value{
A list with two elements: $no_converge, contains codes of all species
with convergence issues, $bad_fit, contains codes of species with goodness-of-fit
issues (small Bayesian p-value).
}
\description{
The datapipeline produces one occupancy diagnostic file for each
species and year. This function combines all these files and returns those
species with convergence or goodnes-of-fit issues.
}
\examples{
\dontrun{
config <- configPipeline(year = 2010,
                         dur = 3,
                         occ_mod = c("log_dist_coast", "elev", "log_hum.km2", "wetcon",
                                     "watrec", "watext", "log_watext", "watext:watrec",
                                     "ndvi", "prcp", "tdiff"),
                         det_mod = c("(1|obs_id)", "log_hours", "prcp", "tdiff", "cwac"),
                         fixed_vars = c("Pentad", "lon", "lat", "watocc_ever", "wetext_2018","wetcon_2018",
                                        "dist_coast", "elev"),
                         package = "spOccupancy",
                         data_dir = "analysis/hpc/imports",
                         out_dir = "analysis/hpc/imports",
                         server = TRUE)
sp_codes <- config$species

selectSppFromDiag(config, sp_codes, 2008)
}
}
