% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selOccuRmod.R
\name{selOccuRmod}
\alias{selOccuRmod}
\title{Select occuR model}
\usage{
selOccuRmod(forms, type, visit_data, site_data, mod_id)
}
\arguments{
\item{forms}{A list with formulas to fit an occuR model see \link[occuR]{fit_occu}}

\item{type}{One of "visit" if the model selection is for the probability of
detection or "site" if model selection is for occupancy probability. An
overall selection to come...}

\item{visit_data}{Occupancy visit data produced by \link{prepDataOccuR} or with
the format prescribed by \link[occuR]{fit_occu}.}

\item{site_data}{Occupancy site data produced by \link{prepDataOccuR} or with
the format prescribed by \link[occuR]{fit_occu}.}

\item{mod_id}{Model identifier. It is recommended to include something that
allows us to identify which model and to what species.
species id.}
}
\value{
A dataframe with four columns: df - estimated degrees of freedom,
form - formula of the assessed model, AIC - AIC score of the model, mod -
model identifier.
}
\description{
Select occuR model
}
