% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{sens_bound}
\alias{sens_bound}
\title{Sensitivity Analysis for BIRDiE Models}
\usage{
sens_bound(x, r_probs, contrast)
}
\arguments{
\item{x}{A \code{birdie} model object.}

\item{r_probs}{A data frame of BISG probabilities, from \code{\link[=bisg]{bisg()}}.}

\item{contrast}{A named numeric vector specifying the estimate, disparity, or
other contrast to compute a sensitivity bound for. Must have length no more
than the number of racial groups. Missing racial groups will have a
coefficient of zero. For example, to get a bound on disparities between
White and Black individuals, use \code{contrast=c(white=1, black=-1, hisp=0, ...)}.}
}
\value{
A vector containing the sensitivity bound factor for the selected
contrast, for each level of the outcome variable. This factor should be
multiplied by the assumed partial \eqn{R^2} of the outcome residuals on
surnames in order to produce a bound on the estimate of the bias
}
\description{
Sensitivity Analysis for BIRDiE Models
}
