% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spOccupancy.R
\name{simDetSpOccu}
\alias{simDetSpOccu}
\title{Simulate detections from spOccupancy fit}
\usage{
simDetSpOccu(object)
}
\arguments{
\item{object}{an spOccupancy fit}
}
\value{
A list with posterior detection probabilities samples and posterior detection
predictions samples. The results are given in a long format and as an attribute the
indices of the sites the samples correspond to
}
\description{
This function comes largely from the fitted.PGOcc.R from the
\href{https://github.com/doserjef/spOccupancy}{spOccupancy} package
}
