% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-occuR.R
\name{summariseOccuR}
\alias{summariseOccuR}
\title{Summarise occuR occupancy estimates}
\usage{
summariseOccuR(pred_p, pred_psi, quants)
}
\arguments{
\item{pred_p}{Detection probabilities estimated from a occuR model.
It must be a matrix with each row corresponding to an bootstrap sample and each column
corresponding to a visit. This object must contain a "pentad" attribute indicating
which pentad each column correspond to, an attribute "year" indicating what
year the probabilities correspond to, and an attribute "obs" indicating whether
the species was detected in the visit or not. Outputs from \code{\link{predictOccuR}},
should be readily appropriate.}

\item{pred_psi}{Occupancy probabilities estimated from an a occuR model.
It must be a matrix with each row corresponding to an bootstrap sample and each column
corresponding to a pentad. This object must contain a "pentad" attribute indicating
which pentad each column correspond to, and an attribute "year" indicating what
year the probabilities correspond to. Outputs from \code{\link{predictOccuR}},
should be readily appropriate.}

\item{quants}{Quantiles to summarise predictions distribution passed as c("lower", "med", "upper").}
}
\value{
A tibble with estimates and/or quantiles for each pentad in site_data:
\itemize{
\item psi: occupancy probability,
\item p: detection probability,
\item occu: realized occupancy (occupancy conditional on data).
}
}
\description{
occuR produces estimates of detection probability for each
visit and occupancy probabilities for each site. This functions takes these
predictions and creates detection probabilities, occupancy probabilities and
realized occupancy probabilities (occupancy probabilities conditional on
observed data) for each site.
}
