% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizePredOccuR.R
\name{summarizePredOccuR}
\alias{summarizePredOccuR}
\title{Summarize occuR predictions for sites}
\usage{
summarizePredOccuR(pred_p, pred_psi, pred_data, visit_data, quants = NULL)
}
\arguments{
\item{pred_p}{Detectopm probabilities estimated from an occuR model. If it is
a matrix it is interpreted as bootstrap predictions, with one row for each
bootstrap sample. The argument 'quants' must then be provided.}

\item{pred_psi}{Occupancy probabilities estimated from an occuR model. If it is
a matrix it is interpreted as bootstrap predictions, with one row for each
bootstrap sample. The argument 'quants' must then be provided.}

\item{pred_data}{A data.frame/data.table with the site data used to generate
occuR model predictions.}

\item{visit_data}{A data.frame/data.table with the visit data used to generate
occuR model predictions.}

\item{quants}{Quantiles to summarize predictions when bootstrap predictions
are supplied in pred_p and pred_psi. Passed as c("lower", "med", "upper").}
}
\value{
A tibble with estimates and/or quantiles for each pentad in site_data:
\itemize{
\item psi: occupancy probability,
\item p: detection probability,
\item occu: realized occupancy (occupancy conditional on data).
}
}
\description{
occuR produces estimates of detection probability for each
visit and occupancy probabilities for each site. This functions takes these
predictions and creates detection probabilities, occupancy probabilities and
realized occupancy probabilities (occupancy probabilities conditional on
observed data) for each site and occasion.
}
