% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jags.R
\name{summaryJags}
\alias{summaryJags}
\title{Summarise JAGS model fit}
\usage{
summaryJags(fit, param = NULL)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{param}{A character vector with the name of the parameters we want to
summarise}
}
\description{
Summarise JAGS model fit
}
