% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbirds.R
\docType{data}
\name{waterbirds}
\alias{waterbirds}
\title{BIRDIE's waterbird masterlist}
\format{
A data frame with 147 rows and 19 variables:
\describe{
\item{item}{DESCRIPTION}
...
}
}
\usage{
waterbirds
}
\description{
A dataset containing information about all of the waterbird species analysed
by the BIRDIE pipeline
}
\keyword{datasets}
