% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJagsModelFile.R
\name{writeJagsModelFile}
\alias{writeJagsModelFile}
\title{Write a JAGS model to a temporary text file}
\usage{
writeJagsModelFile()
}
\value{
This function dumps the model below to a temporary file that calls to rjags can use. It is meant to be a default model that is installed with the package.
}
\description{
Write a JAGS model to a temporary text file
}
\examples{
writeJagsModelFile()
}
