% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdie-package.R
\docType{package}
\name{birdie-package}
\alias{birdie}
\alias{birdie-package}
\title{birdie: Bayesian Instrumental Regression for Disparity Estimation}
\description{
A Bayesian model for accurately estimating conditional distributions by race, using Bayesian Improved Surname Geocoding (BISG) probability estimates of individual race. Implements the methdos described in McCartan, Fisher, Goldin, Ho and Imai (2022) <arXiv:>.
}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

Authors:
\itemize{
  \item Kosuke Imai \email{imai@harvard.edu}
}

}
\keyword{internal}
