% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdie-package.R
\docType{package}
\name{birdie-package}
\alias{birdie-package}
\title{birdie: Bayesian Instrumental Regression for Disparity Estimation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bayesian models for accurately estimating conditional distributions by race, using Bayesian Improved Surname Geocoding (BISG) probability estimates of individual race. Implements the methods described in McCartan, Fisher, Goldin, Ho and Imai (2025) \doi{10.1080/01621459.2025.2526695}.
}
\references{
McCartan, C., Fisher, R., Goldin, J., Ho, D.E., & Imai, K. (2025).
Estimating Racial Disparities when Race is Not Observed. \emph{Journal of the
American Statistical Association}. Available at
\url{https://doi.org/10.1080/01621459.2025.2526695}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CoryMcCartan/birdie}
  \item \url{https://corymccartan.com/birdie/}
  \item Report bugs at \url{https://github.com/CoryMcCartan/birdie/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{mccartan@psu.edu}

Other contributors:
\itemize{
  \item Kosuke Imai \email{imai@harvard.edu} [contributor]
  \item Daniel Ho \email{dho@law.stanford.edu} [contributor]
  \item Jacob Goldin \email{jsgoldin@uchicago.edu} [contributor]
  \item Robin Fisher \email{robin.fisher@treasury.gov} [contributor]
  \item The Stan Development Team (include/rstan) [copyright holder]
}

}
\keyword{internal}
