% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdie.R
\name{birdie.ctrl}
\alias{birdie.ctrl}
\title{Control of BIRDiE Model Fitting}
\usage{
birdie.ctrl(max_iter = 100, abstol = 1e-07, reltol = 1e-06)
}
\arguments{
\item{max_iter}{The maximum number of EM iterations.}

\item{abstol}{The absolute tolerance used in checking convergence.}

\item{reltol}{The relative tolerance used in checking convergence.}
}
\value{
A list containing the control parameters
}
\description{
Constructs control parameters for BIRDiE model fitting.
All arguments have defaults.
}
\examples{
birdie.ctrl(max_iter=1000)

}
