% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdie.R
\name{birdie.ctrl}
\alias{birdie.ctrl}
\title{Control of BIRDiE Model Fitting}
\usage{
birdie.ctrl(
  max_iter = 1000,
  accel = c("squarem", "anderson", "daarem", "none"),
  order = switch(match.arg(accel), none = 0L, anderson = -1L, daarem = -1L, squarem = 3L),
  anderson_restart = TRUE,
  abstol = 1e-06,
  reltol = 1e-06
)
}
\arguments{
\item{max_iter}{The maximum number of EM iterations.}

\item{accel}{The acceleration algorithm to use in doing EM. The default
\code{"squarem"} is good for most purposes, though \code{"anderson"} may be faster
when there are few parameters or very tight tolerances. \code{"daarem"} is an
excellent choice as well that works across a range of problems, though it
requires installing the small \code{SQUAREM} package. \code{"none"} is not
recommended unless other algorithms are running into numerical issues.}

\item{order}{The order to use in the acceleration algorithm. Interpretation
varies by algorithm. Can range from 1 to 3 (default) for SQUAREM and from 1
to the number of parameters for Anderson and DAAREM (default -1 allows the
order to be determined by problem size).}

\item{anderson_restart}{Whether to use restarts in Anderson acceleration.}

\item{abstol}{The absolute tolerance used in checking convergence.}

\item{reltol}{The relative tolerance used in checking convergence.
Ignored if \code{accel="squarem"}.}
}
\value{
A list containing the control parameters
}
\description{
Constructs control parameters for BIRDiE model fitting.
All arguments have defaults.
}
\examples{
birdie.ctrl(max_iter=1000)

}
