% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdie.R
\name{birdie}
\alias{birdie}
\title{Fit BIRDiE Models}
\usage{
birdie(
  r_probs,
  formula,
  data = NULL,
  model = c("auto", "dir", "mmm"),
  prior = NULL,
  prefix = "pr_",
  se_boot = 0,
  ctrl = birdie.ctrl()
)
}
\arguments{
\item{r_probs}{A data frame or matrix of BISG probabilities, with one row per
individual. The output of \code{\link[=bisg]{bisg()}} can be used directly here.}

\item{formula}{A two-sided formula object describing the model structure. The
left-hand side is the outcome variable, which must be discrete. A single
random intercept term, denoted with a vertical bar (\code{"(1 | <term>)"}), is
supported on the right-hand side.}

\item{data}{An optional data frame containing the variables named in \code{formula}.}

\item{model}{A string specifying the type of model to fit: either \code{"dir"} for
the Multinomial-Dirichlet model or \code{"mmm"} for the Multinomial
mixed-effects model. The default, \code{"auto"}, will select the most
computationally efficient model available: \code{"dir"} if \code{formula} has no
covariates or a fully-interacted structure, and \code{"mmm"} otherwise. More
details on the model specifications can be found in the "Details" section
below.}

\item{prior}{A list with entries specifying the model prior. When
\code{model="dir"} the only entry is \code{alpha}, which should be a matrix of
Dirichlet hyperparameters. The matrix should have one row for every level
of the outcome variable and one column for every racial group. The default
prior is a matrix with all entries set to \eqn{1+\epsilon}. When
\code{model="mmm"}, the \code{prior} list should contain two scalar entries:
\code{scale_beta}, the standard deviation on the Normal prior for the fixed
effects, and \code{scale_sigma}, the prior mean of the standard deviation of the
random intercepts.}

\item{prefix}{If \code{r_probs} is a data frame, the columns containing racial
probabilities will be selected as those with names starting with \code{prefix}.
The default will work with the output of \code{\link[=bisg]{bisg()}}.}

\item{se_boot}{The number of bootstrap replicates to use to compute
approximate standard errors for the main model estimates. Only available
when \code{model="dir"}. When there are fewer than 1,000 individuals or 100 or
fewer replicates, a Bayesian bootstrap is used instead (i.e., weights are
drawn from a \eqn{\text{Dirichlet}(1, 1, ..., 1)} distribution, which produces more
reliable estimates.}

\item{ctrl}{A list containing control parameters for the EM algorithm and
optimization routines. A list in the proper format can be made using
\code{\link[=birdie.ctrl]{birdie.ctrl()}}.}
}
\value{
An object of class \code{\link[=birdie-class]{birdie}}, for which many
methods are available. The model estimates may be accessed with
\code{\link[=coef.birdie]{coef.birdie()}}, and updated BISG probabilities (conditioning on the
outcome) may be accessed with \code{\link[=fitted.birdie]{fitted.birdie()}}.
}
\description{
Fits one of two possible Bayesian Instrumental Regression for Disparity
Estimation (BIRDiE) models to BISG probabilities and covariates. The simplest
Multinomial-Dirichlet model (\code{dir}) is appropriate when there are no covariates or when
all covariates are discrete and fully interacted with another. The more
general Multinomial mixed-effects model (\code{mmm}) is a supports any number of
fixed effects and up to one random intercept.
}
\details{
\code{birdie()} uses an expectation-maximization (EM) routine to find the maximum
\emph{a posteriori} (MAP) estimate for the specified model. Asymptotic
variance-covariance matrices for the MAP estimate are available for the
Multinomial-Dirichlet model via bootstrapping (\code{se_boot}).

The Multinomial-Dirichlet model is specified as follows: \deqn{
    Y_i \mid R_i, X_i, \Theta \sim \text{Categorical}(\theta_{R_iX_i}) \\
    \theta_{rx} \sim \text{Dirichlet}(\alpha_r),
} where \eqn{Y} is the outcome variable, \eqn{R} is race, \eqn{X} are
covariates (fixed effects), and \eqn{\theta_{rx}} and \eqn{\alpha_r} are
vectors with length matching the number of levels of the outcome variable.
There is one vector \eqn{\theta_{rx}} for every combination of race and
covariates, hence the need for \code{formula} to either have no covariates or a
fully interacted structure.

The Multinomial mixed-effects model is specified as follows: \deqn{
    Y_i \mid R_i, X_i, \Theta \sim \text{Categorical}(g^{-1}(\mu_{R_iX_i})) \\
    \mu_{rxy} = W\beta_{ry} + Zu_{ry} \\
    u_{ry} \mid \sigma^2_{ry} \sim \mathcal{N}(0, \sigma^2_{ry}) \\
    \beta_{ry} \sim \mathcal{N}(0, s_\beta) \\
    \sigma_{ry} \sim \text{Gamma}(2, 2/s_\sigma),
} where \eqn{\beta_{ry}} are the fixed effects, \eqn{u_{ry}} is the random
intercept, and \eqn{g} is a softmax link function.

More details on the models and their properties may be found in the paper
referenced below.
}
\examples{
data(pseudo_vf)

r_probs = bisg(~ nm(last_name) + zip(zip), data=pseudo_vf)

# Process zip codes to remove missing values
pseudo_vf$zip = proc_zip(pseudo_vf$zip)

birdie(r_probs, turnout ~ 1, data=pseudo_vf)

birdie(r_probs, turnout ~ zip, data=pseudo_vf)

fit = birdie(r_probs, turnout ~ (1 | zip), data=pseudo_vf,
       ctrl=birdie.ctrl(abstol=1e-3))

summary(fit)
coef(fit)
fitted(fit)

}
\references{
McCartan, C., Goldin, J., Ho, D.E., & Imai, K. (2022).
Estimating Racial Disparities when Race is Not Observed.
Available at \url{https://arxiv.org/abs/2303.02580}.
}
\concept{estimators}
