% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{calc_joint_bisgz_ols}
\alias{calc_joint_bisgz_ols}
\title{Estimate the joint distribution of R and X}
\usage{
calc_joint_bisgz_ols(
  r_probs,
  x,
  gz = attr(r_probs, "gz"),
  p_gzr = attr(r_probs, "p_gzr"),
  prefix = "pr_",
  truncate = TRUE
)
}
\arguments{
\item{r_probs}{a matrix data frame of race probabilities}

\item{x}{the variable X to tabulate against}

\item{gz}{values of GZ variables. Usually exported as part of
\code{\link[=predict_race_sgz]{predict_race_sgz()}}}

\item{p_gzr}{a table giving Pr(GZ|R) for use in post-stratification. Columns
should sum to 1.}

\item{prefix}{how to select the race probability columns from \code{r_probs}, if
if is a data frame}

\item{truncate}{if \code{TRUE}, truncate results to lie in [0, 1].}
}
\value{
a matrix
}
\description{
Estimate the joint distribution of R and X
}
