% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configPreambOccuR.R
\name{configPreambOccuR}
\alias{configPreambOccuR}
\title{Prepare server or local preamble}
\usage{
configPreambOccuR(year, dur, dim_grid, server)
}
\arguments{
\item{year}{Year of interest.}

\item{dur}{Temporal coverage of the analysis in years. \code{year} will be the last year
covered by the analysis.}

\item{dim_grid}{An integer giving the dimension of the grid used for spatial effects.
This dimension gives the number \code{k} see \code{\link[mgcv]{choose.k}}.}

\item{server}{Logical. If TRUE the preamble is prepared to run remotely,
otherwise it is prepared to run locally.}
}
\value{
A list with:
\itemize{
\item data_dir: directory where data is retrieved from.
\item fit_dir: directory where fitted model and other output is saved to.
\item dyear: number of years before and after the year of interest.
\item sptemp: spatio-temporal effect of the occupancy model.
\item species: species models will be fitted to.
}
}
\description{
Set basic variables to run BIRDIE scripts locally or remotely.
}
\examples{
configPreambOccuR(year = 2010, dur = 3, dim_grid = 20, server = TRUE)
}
