% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createCombinedExportFile}
\alias{createCombinedExportFile}
\title{Create a combined file for export to data mart}
\usage{
createCombinedExportFile(config, type = c("abu", "dst"))
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPipeline}}

\item{type}{A character string with three options: "abu", for abundance estimates files,
or "dst" for occupancy estimates files.}
}
\value{
It will create a file in \code{config$out_dir/exports} combining all files of the selected
type of all species and years in \code{config$years}.
}
\description{
Create a combined file for export to data mart
}
