% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jags.R
\name{diagnoseGofJagsSsm}
\alias{diagnoseGofJagsSsm}
\title{Diagnose goodness-of-fit for JAGS state-space model}
\usage{
diagnoseGofJagsSsm(fit, counts, config)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}
}
\value{
A data frame with Rhat values for the different parameters estimated
by the model.
}
\description{
This function runs basic posterior predictive checks for a JAGS SSM
}
