% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jags.R
\name{diagnoseRhatJagsSsm}
\alias{diagnoseRhatJagsSsm}
\title{Diagnose convergence for JAGS state-space model}
\usage{
diagnoseRhatJagsSsm(fit, sp_code, config)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{sp_code}{SAFRING reference number of the species we want to analyse.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}
}
\value{
A data frame with Rhat values for the different parameters estimated
by the model.
}
\description{
This function runs basic Rhat checks for a JAGS SSM
}
