% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSsm2ss.R
\name{plotSsm2ss}
\alias{plotSsm2ss}
\title{Plot state series of a CWAC state-space model with two seasons}
\usage{
plotSsm2ss(
  fit,
  ssm_counts,
  linear = TRUE,
  plot_options = list(colors = NULL, pers_theme = NULL)
)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{ssm_counts}{A data frame with the count data use to fit the
state-space model}

\item{linear}{If TRUE (default) abundance estimates and data are
transformed back to its orginal scale.}

\item{plot_options}{A list with two elements: colors - the colours of the
points that will appear in the plot (two values), and pers_theme  - A
personalized ggplot theme.}
}
\value{
A list with two elements: i) plot: a plot with summer and winter fitted states, as well as the long-term trend,
ii) data: the data used to create the individual plots. This is useful for extracting the data used by ggplot to render the plots
(e.g. for exporting to the dashboard)
}
\description{
Plot state series of a CWAC state-space model with two seasons
}
\examples{
\dontrun{
counts <- barberspan
ssmcounts <- prepSsmData(counts, species = NULL)
fit <- fitCwacSsm(ssmcounts, mod_file = "mymodel.jags",
param = c("beta", "lambda", "sig.zeta",
"sig.w", "sig.eps", "sig.alpha", "sig.e", "mu_t", "mu_wt"))
plotSsm2ss(fit = fit, ssm_counts = ssmcounts)
}
}
