% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_create_indtr_file.R
\name{ppl_create_indtr_file}
\alias{ppl_create_indtr_file}
\title{Create indicator storage files}
\usage{
ppl_create_indtr_file(sp_code, overwrite_indtr)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{overwrite_indtr}{Logical. If TRUE, existing files in directories
corresponding to the species in config$species will be overwritten.}
}
\description{
Create indicator storage files
}
