% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_create_site_visit.R
\name{ppl_create_site_visit}
\alias{ppl_create_site_visit}
\title{Create site and visit occupancy files}
\usage{
ppl_create_site_visit(
  config,
  sp_code,
  force_gee_dwld = FALSE,
  force_site_visit = FALSE,
  force_abap_dwld = FALSE,
  monitor_gee = TRUE
)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccu}}

\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{force_gee_dwld}{Whether covariates from Google Earth Engine should be
downloaded, even if a file with covariates is already present on disk.
Defaults to FALSE.}

\item{force_site_visit}{Whether site and visit data should be prepared even if
visit and site data files are already on disk. Defaults to FALSE}

\item{force_abap_dwld}{Indicates whether ABAP data must be downloaded for the species
and years indicated by 'sp_code' and 'years'. If TRUE, data will be downloaded
from ABAP once per session and cached in a temp file. After this the cached
file will be used, unless download is set to FALSE, in which case data will
be downloaded regardless of the cached file.}

\item{monitor_gee}{if TRUE (default) periodic messages of the state of the downloads
from GEE will be printed on screen.}
}
\value{
The first part of the function creates two data frames (in .csv format)
that will be saved to disk: GEE annotated ABAP site data and GEE annotated ABAP
visit data. The second part of the functions creates three data frames that will
be saved to disk: site, visit and species detection data frames,
all in .csv format.
}
\description{
This function prepares site and visit occupancy data to fit an
occupancy model from ABAP data. It has two parts: the first part downloads ABAP
data and annotates them with covariates from Google Earth Engine using the
functions \code{\link{prepGEESiteData()}} and \code{\link{prepGEEVisitData()}},
the second part uses the function \code{\link{createOccuData}} to format the
data.
}
