% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_diagnose_ssm.R
\name{ppl_diagnose_ssm}
\alias{ppl_diagnose_ssm}
\title{Run diagnostics for a state-space model fit}
\usage{
ppl_diagnose_ssm(fit, counts, sp_code, config)
}
\arguments{
\item{fit}{A JAGS state-space model fitted to CWAC data}

\item{counts}{A dataframe with CWAC counts ready for model fit. See \code{\link{ppl_create_data_ssm}}}

\item{sp_code}{SAFRING reference number of the species we want to analyse.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}
}
\value{
A list with Rhat values and posterior check statistics is returned.
At them moment, we obtain Rhat values for all monitored parameters and three
posterior check statistics: "Tmean" proportion of posterior simulations with mean
greater than that observed in the data (we would like values close to 0.5),
"Tsd" proportion of posterior simulations with sd greater than that observed in
the data (we would like values close to 0.5), "Tdiff" mean difference between
observed data and posterior simulations (we would like values close to 0).
}
\description{
This function performs basic checks for an SSM fit, such as
convergence of parameters and posterior predictive checks
}
