% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_estimate_aoo.R
\name{ppl_estimate_aoo}
\alias{ppl_estimate_aoo}
\title{Estimate Area of Occupancy}
\usage{
ppl_estimate_aoo(sp_code, config, verbose, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{verbose}{Logical. If TRUE, the new line added to the indicator table
is displayed in the console.}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Estimate Area of Occupancy
}
