% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_fit_occu_model.R
\name{ppl_fit_occu_model}
\alias{ppl_fit_occu_model}
\title{Fit occupancy model}
\usage{
ppl_fit_occu_model(sp_code, year_sel, config, spatial = FALSE, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year_sel}{Year in data to run model for.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccu}}

\item{spatial}{Whether a spatial model should be fit. Defaults to FALSE.}

\item{...}{Other arguments passed on to other functions}
}
\value{
It can return either a number correspoding to the status of the fitting
process: 1 - No detections, 2 - too few detections, 3 - model fitting failed
(inherited from \code{\link{fitSpOccu}})
}
\description{
Fit occupancy model
}
