% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_fit_ssm_model.R
\name{ppl_fit_ssm_model}
\alias{ppl_fit_ssm_model}
\title{Fit state-space JAGS model}
\usage{
ppl_fit_ssm_model(sp_code, site, year, config, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{site}{Code for site of interest.}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{...}{Other arguments passed on to other functions}
}
\value{

}
\description{
Fit state-space JAGS model
}
