% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_fit_ssm_model.R
\name{ppl_fit_ssm_model}
\alias{ppl_fit_ssm_model}
\title{Fit state-space JAGS model}
\usage{
ppl_fit_ssm_model(counts, sp_code, config, ...)
}
\arguments{
\item{counts}{A data frame with counts generated by \code{\link{ppl_create_data_ssm}}}

\item{sp_code}{SAFRING code of the species of interest}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{...}{Other parameters to pass on to \code{\link{prepGEECatchmData}}}
}
\description{
Fit state-space JAGS model
}
