% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_run_pipe_abu1.R
\name{ppl_run_pipe_abu1}
\alias{ppl_run_pipe_abu1}
\title{Run abundance pipeline ABU1}
\usage{
ppl_run_pipe_abu1(
  sp_code,
  config,
  steps = c("data", "fit", "diagnose", "summary"),
  prep_data_steps,
  ...
)
}
\arguments{
\item{sp_code}{SAFRING reference number of the species we want to analyze.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{steps}{Pipeline steps to run. It can be one or more of: c("data", "fit", "diagnose", "summary").}

\item{prep_data_steps}{Data preparation steps to pass on to \link{ppl_create_data_ssm}}

\item{...}{Other parameters to pass on to \link{prepGEESpCountData}}
}
\description{
Run abundance pipeline ABU1
}
