% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_run_pipe_distr.R
\name{ppl_run_pipe_distr}
\alias{ppl_run_pipe_distr}
\title{Run distribution indicators pipeline}
\usage{
ppl_run_pipe_distr(
  sp_code,
  sp_name,
  year,
  config,
  steps = c("data", "fit", "summ", "indtr"),
  ...
)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{sp_name}{Common name of the species to run the pipeline for. This is
necessary for plots and summaries.}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{steps}{A character vector containing the steps of the pipeline to run.
Can contain:"data", "fit", "summ", "indtr". Defaults to all of them.}

\item{...}{Other arguments passed on to other functions}
}
\description{
Run distribution indicators pipeline
}
