% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_summarise_ssm.R
\name{ppl_summarise_ssm}
\alias{ppl_summarise_ssm}
\title{Summarise predictions from state-space JAGS model}
\usage{
ppl_summarise_ssm(fit, counts, sp_code, config, ...)
}
\arguments{
\item{fit}{A model fit generated by \link{\code{ppl_fit_ssm_model}}}

\item{counts}{A data frame with counts generated by \link{\code{ppl_create_data_ssm}}}

\item{sp_code}{SAFRING code of the species of interest}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{...}{Other parameters to pass on to \link{prepGEESpCountData}}
}
\description{
Summarise predictions from state-space JAGS model
}
