% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl_summarize_ssm.R
\name{ppl_summarize_ssm}
\alias{ppl_summarize_ssm}
\title{Summarize predictions from state-space JAGS model}
\usage{
ppl_summarize_ssm(sp_code, config, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species of interest}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{...}{Other parameters to pass on to \link{prepGEESpCountData}}
}
\description{
Summarize predictions from state-space JAGS model
}
