% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictOccuR.R
\name{predictOccuR}
\alias{predictOccuR}
\title{Predict occupancy from occuR model}
\usage{
predictOccuR(fit, occuRdata, pred_sites, years, scaling = FALSE, boots = 1000)
}
\arguments{
\item{fit}{A occuR model fit. See \link[occuR]{fit_occu}}

\item{occuRdata}{A list produced by \link{prepDataOccuR}}

\item{pred_sites}{A spatial \code{sf} object with the sites to predict occupancy
for.}

\item{years}{A vector with years to predict for.}

\item{scaling}{Logical; whether it is necessary to scale the variables
before predicting. Scaling factors will be extracted from the attributes of
occuRdata. See \link{scale}.}

\item{boots}{Number of bootstrap samples to calculate quantiles.}
}
\value{
A list with predictions for detection and occupancy probabilities.
}
\description{
Predict occupancy from occuR model
}
