% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictOccuR.R
\name{predictOccuR}
\alias{predictOccuR}
\title{Predict from occuR model fit}
\usage{
predictOccuR(sp_code, year, config, ...)
}
\arguments{
\item{sp_code}{SAFRING code of the species to run the pipeline for}

\item{year}{Year to run to the pipeline for}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}

\item{...}{Other arguments passed on to other functions}
}
\description{
Predict from occuR model fit
}
