% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDataOccuR.R
\name{prepDataOccuR}
\alias{prepDataOccuR}
\title{Prepare data for fitting a model with occuR}
\usage{
prepDataOccuR(site_data, visit_data, scaling = NULL)
}
\arguments{
\item{site_data}{A dataframe with occupancy site data.}

\item{visit_data}{A dataframe with occupancy visit data.}

\item{scaling}{Either NULL, in which case no scaling is performed or a list
of two named vectors: "site" and "visit" indicating the variables to scale
in the site and visit dataframes.}
}
\value{
A list containing two data frames: one with site data and one with
visit data ready to use with the occuR package
}
\description{
Prepare data for fitting a model with occuR
}
