% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gee.R
\name{prepGEECatchmData}
\alias{prepGEECatchmData}
\title{Prepare Google Earth Engine data for catchments}
\usage{
prepGEECatchmData(sp_code, catchment, config, monitor = TRUE)
}
\arguments{
\item{sp_code}{SAFRING reference number of the species we want to analyze.}

\item{catchment}{An sf object with the polygons defining the catchments
to be annotated.}

\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{monitor}{Logical. If TRUE (default) monitoring printed messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}
}
\description{
Prepare Google Earth Engine data for catchments
}
\details{
It is assumed that there is an asset on \code{rgee::ee_get_assethome()}
named 'quin_catchm' that has the polygons defining the quinary catchments.
Note that we add an extra year before the start of the series. This is because
summer waterbird populations should be affected by the conditions in the previous
year rather than by conditions in the following year.
}
