% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGEESiteData.R
\name{prepGEESiteData}
\alias{prepGEESiteData}
\title{Prepare Google Earth Engine site data}
\usage{
prepGEESiteData(config)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}
}
\description{
Prepare Google Earth Engine site data
}
