% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGEESiteData.R
\name{prepGEESiteData}
\alias{prepGEESiteData}
\title{Prepare Google Earth Engine site data}
\usage{
prepGEESiteData(config, pentads, asset_id, upload_asset = TRUE, monitor = TRUE)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{pentads}{An sf spatial object with the set of pentads that we want to
annotate with environmental covariates. See \code{\link[ABAP]{getRegionPentads}}}

\item{asset_id}{Character string with the name given to the object created
in Google Earth Engine (asset) that contains the sites in \code{pentads}.}

\item{upload_asset}{If TRUE (default), the object \code{pentads} will be uploaded to Google
Earth Engine and an asset under the name of \code{asset_id} will be created. If FALSE,
it will be assumed that an asset named after \code{asset_id} is already
present in GEE and \code{pentads} will not be uploaded.}

\item{monitor}{Logical. If TRUE (default) monitoring printed messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}
}
\description{
Prepare Google Earth Engine site data
}
