% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGEESpCountData.R
\name{prepGEESpCountData}
\alias{prepGEESpCountData}
\title{Prepare Google Earth Engine data for CWAC species counts}
\usage{
prepGEESpCountData(
  counts,
  sp_code,
  catchment,
  config,
  upload_catchment = FALSE,
  force_gee = FALSE
)
}
\arguments{
\item{counts}{A dataframe with counts to annotate with covariates. This data
frame must contain at least the lat-lon coordinates of the CWAC site the
counts were obtained at. The variables must be named "X" and "Y".}

\item{sp_code}{SAFRING reference number of the species we want to analyze.}

\item{catchment}{An sf object with the polygons defining the catchments
of the CWAC sites in \code{counts}. Which site belongs to which catchment will be
determined by a join in GEE.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambJAGS}}

\item{upload_catchment}{If TRUE catchment will be uploaded to GEE. FALSE is
an option because catchment info might have been already uploaded to GEE.
Defaults to FALSE.}

\item{force_gee}{If TRUE annotation with GEE info will occur even if a
dataset with covariates is already present on disk.}
}
\value{

}
\description{
Prepare Google Earth Engine data for CWAC species counts
}
