% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGEEVisitData.R
\name{prepGEEVisitData}
\alias{prepGEEVisitData}
\title{Prepare Google Earth Engine visit data}
\usage{
prepGEEVisitData(config)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccuR}}
}
\value{

}
\description{
Prepare Google Earth Engine visit data
}
