% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepGEEVisitData.R
\name{prepGEEVisitData}
\alias{prepGEEVisitData}
\title{Prepare Google Earth Engine visit data}
\usage{
prepGEEVisitData(config, visits, asset_id, upload_asset = TRUE, monitor = TRUE)
}
\arguments{
\item{config}{A list with pipeline configuration parameters.
See \code{\link{configPipeline}}}

\item{visits}{An sf spatial object with the visit data that we want to
annotate with environmental covariates. Note that we need these data to be
in a spatial format, so we should probably join them with pentad data.
See \code{\link[ABAP]{getRegionPentads}}}

\item{asset_id}{Character string with the name given to the object created
in Google Earth Engine (asset) that contains the sites in \code{visits}.}

\item{upload_asset}{If TRUE (default), the object \code{visits} will be uploaded to Google
Earth Engine and an asset under the name of \code{asset_id} will be created. If FALSE,
it will be assumed that an asset named after \code{asset_id} is already present in GEE
and \code{visits} will not be uploaded.}

\item{monitor}{Logical. If TRUE (default) monitoring printed messages produced
by \code{rgee} will displayed. If FALSE, only high-level messages will be displayed.}
}
\description{
Prepare Google Earth Engine visit data
}
\details{
Note that some GEE layers don't have information past a certain date. At the time
of writing surface water layers only have information up until 2021 and human
population density up until 2020. We have set up the code in such a way that
visits past the last date of the layer get annotated with the latest available
information. Take this into consideration for the analyses. Code should be updated
as more information becomes available.
}
