% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepOccuRData.R
\name{prepOccuRData}
\alias{prepOccuRData}
\title{Prepare occuR data for model fitting}
\usage{
prepOccuRData(site_data, visit_data, config, spatial = FALSE, sp_sites = NULL)
}
\arguments{
\item{site_data}{A data frame containing information about covariates associated
with ABAP pentads for a given year.}

\item{visit_data}{A data frame with information associated with sampling visits
to ABAP pentads in a given year. Detection/non-detection data for the species
of interest must also be included in this data frame.}

\item{config}{A list with pipeline configuration parameters.
See \link{configPreambOccu}}

\item{spatial}{Logical, indicating whether spatial random effects should be
included in the model. Defaults to FALSE.}
}
\description{
Prepare occuR data for model fitting
}
