% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSsmData.R
\name{prepSsmData}
\alias{prepSsmData}
\title{Prepare CWAC data to fit a state-space model}
\usage{
prepSsmData(counts, spp_sel = NULL, keep = NULL)
}
\arguments{
\item{counts}{A data frame with raw CWAC count data}

\item{spp_sel}{An optional vector of species codes. Only records for these
species are prepared for fitting an SSM, others are discarded.}

\item{keep}{A vector of variables to keep after the processing. Useful, if
there are covariates of interest. If NULL, only year, season, start date,
count and species name are returned.}
}
\value{
A tibble with clean and prepared data for fitting an SSM (e.g. filled with missing years)
}
\description{
Prepare CWAC data to fit a state-space model
}
\examples{
counts <- barberspan
prepSsmData(counts)
prepSsmData(counts, spp_sel = 212)
prepSsmData(counts, spp_sel = c(212, 50))
}
