% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{rake}
\alias{rake}
\title{Rake a joint distribution so that the margins match provided vectors}
\usage{
rake(est, row = rowSums(est), col = colSums(est), iter = 5)
}
\arguments{
\item{est}{the starting joint distribution}

\item{row}{the desired row sums / margins}

\item{col}{the desired column sums / margins}

\item{iter}{the number of raking iterations}
}
\value{
a new matrix \code{est}
}
\description{
Rake a joint distribution so that the margins match provided vectors
}
\keyword{internal}
